Compare EGM with EFA

Estimates an EGM based on EGA and uses the number of communities as the number of dimensions in exploratory factor analysis (EFA) using fa

Usage

EGM.compare(
  data,
  constrain.structure = FALSE,
  constrain.zeros = FALSE,
  fm = "ml",
  rotation = "geominQ",
  ...
)

Arguments

data

Matrix or data frame. Should consist only of variables to be used in the analysis. Can be raw data or a correlation matrix

constrain.structure

Boolean (length = 1). Whether memberships of the communities should be added as a constraint when optimizing the network loadings. Defaults to TRUE which ensures assigned loadings are guaranteed to never be smaller than any cross-loadings. Set to FALSE to freely estimate each loading similar to exploratory factor analysis

Note: This default differs from EGM. Constraining loadings puts EGM at a deficit relative to EFA and therefore biases the comparability between the methods. It's best to leave the default of unconstrained when using this function.

constrain.zeros

Boolean (length = 1). Whether zeros in the estimated network loading matrix should be retained when optimizing the network loadings. Defaults to TRUE which ensures that zero networks loadings are retained. Set to FALSE to freely estimate each loading similar to exploratory factor analysis

Note: This default differs from EGM. Constraining zeros puts EGM at a deficit relative to EFA and therefore biases the comparability between the methods. It's best to leave the default of unconstrained when using this function.

fm

Character (length = 1). Estimation method for the EFA. See argument in fa for more details. Defaults to "ml" or maximum likelihood

rotation

Character (length = 1). A rotation to use to obtain a simpler structure for EFA. For a list of rotations, see rotations for options. Defaults to "geominQ"

...

Additional arguments to be passed on to auto.correlate, network.estimation, community.detection, community.consensus, community.unidimensional, EGA, EGM, net.loads, and fa

Author

Hudson F. Golino <hfg9s at virginia.edu> and Alexander P. Christensen <alexpaulchristensen@gmail.com>

Examples

# Get depression data
data <- depression[,24:44]

# Compare EGM (using EGA) with EFA
if (FALSE) { # \dontrun{
results <- EGM.compare(data)

# Print summary
summary(results)} # }