Unique Variable Analysis

Identifies locally dependent (redundant) variables in a multivariate dataset using the EBICglasso.qgraph network estimation method and weighted topological overlap (see Christensen, Garrido, & Golino, 2023 for more details)

Usage

UVA(
  data = NULL,
  network = NULL,
  n = NULL,
  key = NULL,
  uva.method = c("MBR", "EJP"),
  cut.off = 0.25,
  reduce = TRUE,
  reduce.method = c("latent", "mean", "remove", "sum"),
  auto = TRUE,
  verbose = FALSE,
  ...
)

Arguments

data

Matrix or data frame. Should consist only of variables to be used in the analysis. Can be raw data or a correlation matrix. Defaults to NULL

network

Symmetric matrix or data frame. A symmetric network. Defaults to NULL

If both data and network are provided, then UVA will use the network with the data (rather than estimating a network from the data)

n

Numeric (length = 1). Sample size if data provided is a correlation matrix. Defaults to NULL

key

Character vector (length = ncol(data)). Item key for labeling variables in the results

uva.method

Character (length = 1). Whether the method described in Christensen, Garrido, and Golino (2023) publication in Multivariate Behavioral Research ("MBR") or Christensen, Golino, and Silvia (2020) publication in European Journal of Personality ("EJP") should be used. Defaults to "MBR"

Based on simulation and accumulating empirical evidence, the methods described in Christensen, Golino, and Silvia (2020) such as adaptive alpha are outdated. Evidence supports using a single cut-off value (regardless of continuous, polytomous, or dichotomous data; Christensen, Garrido, & Golino, 2023)

cut.off

Numeric (length = 1). Cut-off used to determine when pairwise wto values are considered locally dependent (or redundant). Must be values between 0 and 1. Defaults to 0.25

This cut-off value is recommended and based on extensive simulation (Christensen, Garrido, & Golino, 2023). Printing the result will provide a gradient of pairwise redundancies in increments of 0.20, 0.25, and 0.30. Use print or summary on the output rather than adjusting this cut-off value

reduce

Logical (length = 1). Whether redundancies should be reduced in data. Defaults to TRUE

reduce.method

Character (length = 1). Method to reduce redundancies. Available options:

"latent" — Computes latent variables using cfa when there are three or more redundant variables. If variables are not all coded in the same direction, then they will be recoded as necessary. A warning will be produced for all variables that are flipped

"mean" — Computes mean of redundant variables. If variables are not all coded in the same direction, then they will be recoded as necessary. A warning will be produced for all variables that are flipped

"remove" — Removes all but one variable from a set of redundant variables

"sum" — Computes sum of redundant variables. If variables are not all coded in the same direction, then they will be recoded as necessary. A warning will be produced for all variables that are flipped

auto

Logical (length = 1). Whether reduce should occur automatically. For reduce.method = "remove", the automated decision process is as follows:

Two variables — The variable with the lowest maximum wto to all other variables (other than the one it is redundant with) is retained and the other is removed

Three or more variables — The variable with the highest mean wto to all other variables that are redundant with one another is retained and all others are removed

verbose

Boolean (length = 1). Whether messages and (insignificant) warnings should be output. Defaults to FALSE (silent calls). Set to TRUE to see all messages and warnings for every function call

...

Additional arguments that should be passed on to old versions of UVA or to EGA and cfa

References

Most recent simulation and implementation
Christensen, A. P., Garrido, L. E., & Golino, H. (2023). Unique variable analysis: A network psychometrics method to detect local dependence. Multivariate Behavioral Research.

Conceptual foundation and outdated methods
Christensen, A. P., Golino, H., & Silvia, P. J. (2020). A psychometric network perspective on the validity and validation of personality trait questionnaires. European Journal of Personality, 34(6), 1095-1108.

Weighted topological overlap
Nowick, K., Gernat, T., Almaas, E., & Stubbs, L. (2009). Differences in human and chimpanzee gene expression patterns define an evolving network of transcription factors in brain. Proceedings of the National Academy of Sciences, 106, 22358-22363.

Selection of CFA Estimator
Rhemtulla, M., Brosseau-Liard, P. E., & Savalei, V. (2012). When can categorical variables be treated as continuous? A comparison of robust continuous and categorical SEM estimation methods under suboptimal conditions. Psychological Methods, 17(3), 354-373.

Examples

# Perform UVA
uva.wmt <- UVA(wmt2[,7:24])

# Show summary
summary(uva.wmt)
#> Variable pairs with wTO > 0.30 (large-to-very large redundancy)
#> 
#> ----
#> 
#> Variable pairs with wTO > 0.25 (moderate-to-large redundancy)
#> 
#> ----
#> 
#> Variable pairs with wTO > 0.20 (small-to-moderate redundancy)
#> 
#>  node_i node_j  wto
#>    wmt2   wmt3 0.25