Ergodicity Information Index

Computes the Ergodicity Information Index

Usage

ergoInfo(
  dynEGA.object,
  use = c("edge.list", "unweighted"),
  ordering = c("row", "column")
)

Arguments

dynEGA.object

A dynEGA.ind.pop object

use

Character (length = 1). A string indicating what network element will be used to compute the algorithm complexity, the list of edges or the weights of the network. Defaults to use = "edge.list". Current options are:

"edge.list" --- Calculates the algorithm complexity using the list of edges.

"unweighted" --- Calculates the algorithm complexity using the binary weights of the network. 0 = edge absent and 1 = edge present

ordering

Character (length = 1). Changes ordering of edge list. "row" goes across the rows; "column" goes down the columns. Defaults to "column" (original implementation)

Value

Returns a list containing:

PrimeWeight

The prime-weight encoding of the individual networks

PrimeWeight.pop

The prime-weight encoding of the population network

Kcomp

The Kolmogorov complexity of the prime-weight encoded individual networks

Kcomp.pop

The Kolmogorov complexity of the prime-weight encoded population network

complexity

The complexity metric proposed by Santora and Nicosia (2020)

EII

The Ergodicity Information Index

Author

Hudson Golino <hfg9s at virginia.edu> and Alexander Christensen <alexpaulchristensen@gmail.com>

Examples

# Obtain data
sim.dynEGA <- sim.dynEGA # bypasses CRAN checks

if (FALSE) {
# Dynamic EGA individual and population structure
dyn.ega1 <- dynEGA.ind.pop(
  data = sim.dynEGA[,-26], n.embed = 5, tau = 1,
  delta = 1, id = 25, use.derivatives = 1,
  ncores = 2, corr = "pearson"
)

# Compute empirical ergodicity information index
eii <- ergoInfo(
  dynEGA.object = dyn.ega1,
  use = "edge.list"
)}