Network Scores

This function computes network scores computed based on each node's strength within each community in the network (see net.loads). These values are used as "network loadings" for the weights of each variable.

Network scores are computed as a formative composite rather than a reflective factor. This composite representation is consistent with no latent factors that psychometric network theory proposes.

Scores can be computed as a "simple" structure, which is equivalent to a weighted sum scores or as a "full" structure, which is equivalent to an EFA approach. Conservatively, the "simple" structure approach is recommended until further validation

Usage

net.scores(
  data,
  A,
  wc,
  loading.method = c("original", "revised"),
  rotation = NULL,
  scores = c("Anderson", "Bartlett", "components", "Harman", "network", "tenBerge",
    "Thurstone"),
  loading.structure = c("simple", "full"),
  impute = c("mean", "median", "none"),
  ...
)

Arguments

data

Matrix or data frame. Should consist only of variables to be used in the analysis

A

Network matrix, data frame, or EGA object

wc

Numeric or character vector (length = ncol(A)). A vector of community assignments. If input into A is an EGA object, then wc is automatically detected

loading.method

Character (length = 1). Sets network loading calculation based on implementation described in "original" (Christensen & Golino, 2021) or the "revised" (Christensen et al., 2024) implementation. Defaults to "revised"

rotation

Character. A rotation to use to obtain a simpler structure. For a list of rotations, see rotations for options. Defaults to NULL or no rotation. By setting a rotation, scores estimation will be based on the rotated loadings rather than unrotated loadings

scores

Character (length = 1). How should scores be estimated? Defaults to "network" for network scores. Set to other scoring methods which will be computed using factor.scores (see link for arguments and explanations for other methods)

loading.structure

Character (length = 1). Whether simple structure or the saturated loading matrix should be used when computing scores. Defaults to "simple"

"simple" structure more closely mirrors sum scores and CFA; "full" structure more closely mirrors EFA

Simple structure is the more "conservative" (established) approach and is therefore the default. Treat "full" as experimental as proper vetting and validation has not been established

impute

Character (length = 1). If there are any missing data, then imputation can be implemented. Available options:

"none" — Default. No imputation is performed

"mean" — The mean value of each variable is used to replace missing data for that variable

"median" — The median value of each variable is used to replace missing data for that variable

...

Additional arguments to be passed on to net.loads and factor.scores

Value

Returns a list containing:

scores

A list containing the standardized (std.scores) rotated (rot.scores) scores. If rotation = NULL, then rot.scores will be NULL

loadings

Output from net.loads

References

Original implementation and simulation for loadings
Christensen, A. P., & Golino, H. (2021). On the equivalency of factor and network loadings. Behavior Research Methods, 53, 1563-1580.

Preliminary simulation for scores
Golino, H., Christensen, A. P., Moulder, R., Kim, S., & Boker, S. M. (2021). Modeling latent topics in social media using Dynamic Exploratory Graph Analysis: The case of the right-wing and left-wing trolls in the 2016 US elections. Psychometrika.

Revised network loadings
Christensen, A. P., Golino, H., Abad, F. J., & Garrido, L. E. (2024). Revised network loadings. PsyArXiv.

Author

Alexander P. Christensen <alexpaulchristensen@gmail.com> and Hudson F. Golino <hfg9s at virginia.edu>

Examples

# Load data
wmt <- wmt2[,7:24]

# Estimate EGA
ega.wmt <- EGA(
  data = wmt,
  plot.EGA = FALSE # No plot for CRAN checks
)

# Network scores
net.scores(data = wmt, A = ega.wmt)
#> The default 'loading.method' has changed to "revised" in {EGAnet} version >= 2.0.7.
#> 
#>  For the previous default (version <= 2.0.6), use `loading.method = "original"`
#> $scores
#> $scores$std.scores
#>                   1            2
#>    [1,] -1.16628719  0.306711594
#>    [2,]  1.24929277  4.595004206
#>    [3,]  0.64929550  1.915813506
#>    [4,] -0.89839873 -2.844776087
#>    [5,]  1.24929277 -2.009917522
#>    [6,]  1.24929277 -3.598234276
#>    [7,] -2.63724729 -1.166915815
#>    [8,]  1.24929277  5.926400834
#>    [9,]  0.56572589 -0.177986940
#>   [10,]  0.64929550 -1.834083359
#>   [11,]  0.56572589  0.763026420
#>   [12,]  0.64929550  3.878006924
#>   [13,]  0.38516542 -1.049054531
#>   [14,]  1.24929277 -1.967116629
#>   [15,]  1.24929277 -2.336866525
#>   [16,] -2.44985134 -3.791324470
#>   [17,]  1.24929277 -2.111110435
#>   [18,] -0.29840146 -2.609305401
#>   [19,] -1.58572399 -2.886891712
#>   [20,]  1.24929277  4.467294796
#>   [21,]  1.24929277  5.926400834
#>   [22,] -0.21483185  1.982118854
#>   [23,]  1.24929277  2.239409336
#>   [24,] -1.68579195  0.261983266
#>   [25,]  1.24929277  1.223399199
#>   [26,] -1.76628446 -2.240069270
#>   [27,]  1.24929277 -3.779975038
#>   [28,] -3.23724456 -4.533433226
#>   [29,] -0.03427138 -0.269931926
#>   [30,]  1.24929277  5.310344584
#>   [31,]  1.24929277  4.131231843
#>   [32,] -0.29840146 -1.873805249
#>   [33,]  1.24929277  4.721056713
#>   [34,]  1.24929277 -2.886891712
#>   [35,]  1.24929277  1.515446824
#>   [36,]  0.46189956 -0.659006253
#>   [37,]  0.64929550  0.080703622
#>   [38,] -0.82166460 -3.031658998
#>   [39,]  0.64929550  2.168228285
#>   [40,] -3.23724456  0.799487327
#>   [41,]  1.24929277  2.987725376
#>   [42,] -0.82166460 -1.384420624
#>   [43,] -0.03427138 -0.358057070
#>   [44,]  1.24929277  0.544475985
#>   [45,] -0.13809771 -3.480372992
#>   [46,]  1.24929277  2.367118746
#>   [47,] -0.21483185 -1.971233417
#>   [48,]  1.24929277  1.282510306
#>   [49,]  0.64929550  5.184292077
#>   [50,] -1.58572399 -0.898632527
#>   [51,]  1.24929277  3.644316968
#>   [52,]  0.64929550 -1.681547591
#>   [53,] -1.68579195 -3.062977756
#>   [54,] -0.22166733 -1.373076114
#>   [55,]  0.64929550  1.238998327
#>   [56,] -0.03427138 -0.027564936
#>   [57,]  0.46189956 -3.598234276
#>   [58,] -0.82166460 -0.497682207
#>   [59,]  1.24929277  1.779140193
#>   [60,]  1.24929277  0.472979279
#>   [61,]  1.24929277  5.926400834
#>   [62,]  0.38516542  5.172942645
#>   [63,] -0.03427138  2.188971091
#>   [64,]  0.56572589  3.531717151
#>   [65,]  0.38516542  1.683728235
#>   [66,]  1.24929277  2.900958787
#>   [67,] -3.23724456 -3.662365635
#>   [68,]  0.64929550  4.370692504
#>   [69,]  0.64929550  0.544475985
#>   [70,] -0.21483185 -2.727166684
#>   [71,]  0.64929550  3.414234562
#>   [72,]  1.24929277  5.033317508
#>   [73,] -0.03427138 -1.501482383
#>   [74,]  1.24929277 -0.491337299
#>   [75,]  1.24929277  0.313002370
#>   [76,] -1.08955305 -3.706380963
#>   [77,] -0.89839873 -3.640349901
#>   [78,]  0.38516542  0.502200669
#>   [79,]  1.24929277  1.109394359
#>   [80,] -1.76628446  0.434421004
#>   [81,]  1.24929277  1.582477851
#>   [82,]  1.24929277 -0.866004451
#>   [83,]  0.46189956 -0.152322523
#>   [84,] -1.00222507 -4.533433226
#>   [85,]  0.38516542  0.520106488
#>   [86,]  0.64929550 -1.250119348
#>   [87,] -1.16628719  0.633641456
#>   [88,]  1.24929277  3.745258000
#>   [89,]  0.64929550  3.856969885
#>   [90,]  1.24929277  4.253167562
#>   [91,]  0.38516542 -0.208982203
#>   [92,] -0.22166733  2.157652333
#>   [93,] -0.21483185  1.474209292
#>   [94,] -0.21483185 -2.609305401
#>   [95,]  1.24929277  1.391174079
#>   [96,] -2.37311720 -2.363634903
#>   [97,]  1.24929277  2.200320524
#>   [98,]  0.56572589 -1.212120510
#>   [99,]  0.64929550  2.312920340
#>  [100,]  0.38516542  2.276459433
#>  [101,]  1.24929277  3.680463601
#>  [102,]  0.56572589  0.249424320
#>  [103,]  0.64929550 -2.037264479
#>  [104,]  1.24929277  1.882079897
#>  [105,]  0.38516542  2.090265542
#>  [106,]  1.24929277  4.595004206
#>  [107,]  1.24929277 -1.116834197
#>  [108,] -0.98572672  3.191579763
#>  [109,]  1.24929277  4.471369231
#>  [110,]  1.24929277  4.357239814
#>  [111,]  0.56572589  1.735298736
#>  [112,]  1.24929277  5.337112962
#>  [113,]  1.24929277  4.873340599
#>  [114,]  0.38516542  3.767951943
#>  [115,] -0.03427138 -2.110858554
#>  [116,]  1.24929277  5.926400834
#>  [117,]  0.64929550  2.438972847
#>  [118,]  0.46189956  2.115377017
#>  [119,] -3.23724456 -0.320015879
#>  [120,]  1.24929277  2.115377017
#>  [121,]  1.24929277  3.329291055
#>  [122,]  1.24929277  2.024619063
#>  [123,]  1.24929277  4.568235828
#>  [124,] -0.98572672  0.720751104
#>  [125,]  1.24929277  2.364527500
#>  [126,]  1.24929277  3.978947956
#>  [127,] -0.21483185  1.373268261
#>  [128,]  1.24929277  4.471369231
#>  [129,]  1.24929277  2.902208212
#>  [130,]  1.24929277  4.572310263
#>  [131,]  0.64929550 -2.115863079
#>  [132,]  1.24929277  5.188366512
#>  [133,]  0.46189956  0.995264943
#>  [134,] -0.89839873 -0.322390358
#>  [135,]  1.24929277  2.896884352
#>  [136,]  0.56572589 -0.282576278
#>  [137,]  1.24929277  0.003753822
#>  [138,] -0.13809771  4.525567637
#>  [139,]  1.24929277 -0.309344656
#>  [140,]  0.64929550  1.373268261
#>  [141,] -0.89839873 -3.791324470
#>  [142,]  0.38516542 -0.432998282
#>  [143,]  0.56572589  0.552030406
#>  [144,]  0.38516542  2.544514108
#>  [145,]  0.64929550  1.441852749
#>  [146,]  1.24929277  0.606397132
#>  [147,]  0.38516542 -0.113855580
#>  [148,] -0.03427138 -3.917376977
#>  [149,]  1.24929277  0.453416557
#>  [150,]  0.38516542  5.099348571
#>  [151,]  1.24929277  2.115377017
#>  [152,]  1.24929277  4.599078641
#>  [153,]  1.24929277  2.242136634
#>  [154,]  1.24929277  2.413861372
#>  [155,]  1.24929277  4.244639998
#>  [156,] -0.21483185  3.266025110
#>  [157,]  1.24929277  4.568235828
#>  [158,]  1.24929277 -3.598234276
#>  [159,]  0.64929550 -1.370348816
#>  [160,]  1.24929277  1.528143393
#>  [161,]  1.24929277  1.302484554
#>  [162,] -0.29840146 -1.362404890
#>  [163,]  1.24929277  2.268905012
#>  [164,]  1.24929277  4.467294796
#>  [165,] -3.23724456 -1.900114421
#>  [166,]  1.24929277  4.361314249
#>  [167,] -1.00222507 -0.193410807
#>  [168,]  1.24929277  3.856969885
#>  [169,]  1.24929277  1.368341745
#>  [170,] -0.13809771  1.091488541
#>  [171,] -0.03427138  0.472979279
#>  [172,] -0.22166733 -2.609305401
#>  [173,]  0.56572589  0.607594830
#>  [174,]  0.64929550  0.983341853
#>  [175,]  0.64929550 -0.249119779
#>  [176,] -0.03427138  1.915813506
#>  [177,]  1.24929277  0.438069310
#>  [178,]  1.24929277  0.940640785
#>  [179,]  1.24929277 -4.533433226
#>  [180,] -1.00222507 -3.640349901
#>  [181,]  0.64929550  3.575190165
#>  [182,]  0.64929550  1.678323400
#>  [183,] -0.98572672  3.619205493
#>  [184,]  0.38516542 -1.166915815
#>  [185,] -0.21483185  4.439276993
#>  [186,] -1.00222507 -2.270835463
#>  [187,]  1.24929277  3.506605676
#>  [188,]  1.24929277  3.701242672
#>  [189,]  1.24929277 -0.023111807
#>  [190,]  0.38516542 -0.164966875
#>  [191,]  1.24929277 -3.008946404
#>  [192,]  0.56572589  1.192429573
#>  [193,] -0.21483185 -3.051062030
#>  [194,]  1.24929277  5.926400834
#>  [195,]  1.24929277  3.741183565
#>  [196,]  0.64929550  2.179887070
#>  [197,]  1.24929277  5.188366512
#>  [198,]  0.38516542 -1.353101866
#>  [199,] -0.98572672  0.046029138
#>  [200,] -0.82166460 -1.502259841
#>  [201,] -0.40222780 -2.840322958
#>  [202,]  1.24929277  5.184292077
#>  [203,]  0.38516542  0.742592967
#>  [204,] -1.00222507 -0.118398121
#>  [205,]  1.24929277  3.575190165
#>  [206,] -0.89839873  1.944001044
#>  [207,]  0.64929550  2.517348385
#>  [208,] -0.21483185  0.475432291
#>  [209,] -0.03427138 -3.024293651
#>  [210,] -0.29840146 -3.598234276
#>  [211,]  1.24929277  4.599078641
#>  [212,]  1.24929277  2.635538085
#>  [213,] -0.89839873 -0.657659115
#>  [214,]  1.24929277  1.309941724
#>  [215,]  1.24929277  2.804084911
#>  [216,]  0.46189956 -1.828176760
#>  [217,]  1.24929277  4.620115681
#>  [218,] -3.23724456 -3.662365635
#>  [219,] -0.13809771 -0.889169813
#>  [220,]  1.24929277 -0.339863552
#>  [221,]  1.24929277 -0.253572907
#>  [222,] -1.08955305  0.364558306
#>  [223,]  1.24929277  5.188366512
#>  [224,] -0.29840146 -1.528726707
#>  [225,]  1.24929277  0.310912842
#>  [226,]  0.56572589 -0.295848223
#>  [227,]  1.24929277 -2.705150950
#>  [228,] -0.82166460 -0.320015879
#>  [229,]  1.24929277  0.438069310
#>  [230,]  1.24929277  3.619205493
#>  [231,]  1.24929277  3.619205493
#>  [232,]  1.24929277  2.828299863
#>  [233,]  1.24929277  1.474209292
#>  [234,] -0.13809771  3.499275781
#>  [235,]  0.64929550  4.986748754
#>  [236,] -0.03427138 -1.716222076
#>  [237,]  1.24929277  5.209403552
#>  [238,] -0.89839873 -1.947239633
#>  [239,] -2.55367767 -1.747578654
#>  [240,]  1.24929277 -0.403994911
#>  [241,] -3.23724456 -3.008946404
#>  [242,]  0.64929550  1.309941724
#>  [243,]  0.38516542  1.864257571
#>  [244,]  1.24929277  2.200320524
#>  [245,]  1.24929277 -2.224009767
#>  [246,]  1.24929277  2.938354845
#>  [247,] -0.30215984  4.552336015
#>  [248,]  1.24929277  4.317298921
#>  [249,] -0.90215711  2.858192884
#>  [250,]  1.24929277  5.926400834
#>  [251,] -0.30215984  3.607856061
#>  [252,]  0.64929550 -2.168145828
#>  [253,]  1.24929277  4.446257756
#>  [254,] -0.82166460 -1.986405065
#>  [255,] -0.40222780  0.326114626
#>  [256,]  1.24929277  0.326274316
#>  [257,]  1.24929277  3.619205493
#>  [258,]  0.64929550  1.197182217
#>  [259,] -0.90215711  1.073925936
#>  [260,] -1.00222507  2.709118017
#>  [261,]  0.64929550  4.467294796
#>  [262,] -2.55367767  3.767951943
#>  [263,] -0.13809771 -3.640349901
#>  [264,] -0.03427138 -1.470551578
#>  [265,]  0.64929550 -1.203376722
#>  [266,]  0.64929550  2.023382399
#>  [267,]  1.24929277  3.619205493
#>  [268,]  0.38516542 -1.920374003
#>  [269,] -0.13809771  1.148782687
#>  [270,]  1.24929277  2.890549427
#>  [271,]  1.24929277  3.531717151
#>  [272,]  1.24929277  4.249093127
#>  [273,] -0.03427138 -1.275062502
#>  [274,]  0.38516542 -0.333035944
#>  [275,] -1.00222507  2.736788470
#>  [276,]  1.24929277  3.882081359
#>  [277,] -0.40222780  1.180178066
#>  [278,] -1.58572399 -3.917376977
#>  [279,]  0.64929550  1.779314064
#>  [280,]  0.38516542  1.972739210
#>  [281,]  1.24929277  1.516484608
#>  [282,]  1.24929277  5.188366512
#>  [283,] -1.00222507 -1.834083359
#>  [284,]  1.24929277  3.793063418
#>  [285,] -0.30215984  5.926400834
#>  [286,]  1.24929277  3.793063418
#>  [287,]  0.38516542  0.240680531
#>  [288,]  1.24929277 -3.816435945
#>  [289,]  0.64929550  1.277044663
#>  [290,] -3.23724456 -2.462345374
#>  [291,] -0.89839873 -2.169894431
#>  [292,] -0.89839873 -1.166915815
#>  [293,] -0.89839873  1.962556132
#>  [294,] -1.84985407 -1.398865797
#>  [295,]  0.64929550  4.446257756
#>  [296,]  1.24929277 -0.093595999
#>  [297,]  1.24929277 -3.816435945
#>  [298,]  0.46189956 -1.663801463
#>  [299,] -0.03427138 -3.640349901
#>  [300,]  1.24929277  0.198405215
#>  [301,]  0.64929550  3.155970128
#>  [302,]  1.24929277  2.613906934
#>  [303,]  1.24929277 -3.598234276
#>  [304,]  1.24929277 -0.854655019
#>  [305,]  0.56572589  2.355195656
#>  [306,]  1.24929277  2.917317805
#>  [307,] -0.29840146 -2.421810032
#>  [308,]  0.46189956 -1.760397094
#>  [309,]  1.24929277  4.595004206
#>  [310,]  0.38516542 -2.219257123
#>  [311,]  1.24929277  1.980461951
#>  [312,] -0.89839873 -1.834083359
#>  [313,]  1.24929277  5.188366512
#>  [314,] -3.23724456 -1.787514605
#>  [315,]  0.64929550 -3.779975038
#>  [316,]  0.64929550 -1.834083359
#>  [317,]  0.38516542  2.413861372
#>  [318,]  0.64929550  0.603141701
#>  [319,] -0.29840146 -0.424547338
#>  [320,]  1.24929277  5.337112962
#>  [321,]  1.24929277  5.055333242
#>  [322,]  0.38516542 -1.049054531
#>  [323,]  1.24929277  1.520778047
#>  [324,] -0.22166733 -0.588299167
#>  [325,]  1.24929277  0.326274316
#>  [326,]  0.38516542  2.801531485
#>  [327,] -3.23724456 -1.798612156
#>  [328,]  1.24929277 -2.270835463
#>  [329,]  1.24929277  0.402245205
#>  [330,] -1.77311993 -2.908907447
#>  [331,]  0.46189956  5.188366512
#>  [332,]  0.38516542  0.428689665
#>  [333,]  0.38516542 -1.963042194
#>  [334,]  1.24929277  2.679553413
#>  [335,] -0.29840146 -1.343001858
#>  [336,] -0.03427138 -3.593781147
#>  [337,] -3.23724456 -3.227148073
#>  [338,] -0.29840146  0.825796499
#>  [339,] -0.03427138 -2.473689885
#>  [340,] -1.68579195 -2.769282310
#>  [341,]  1.24929277  0.710469385
#>  [342,]  0.64929550 -1.339746427
#>  [343,]  1.24929277  0.757212011
#>  [344,]  0.64929550  2.046250214
#>  [345,] -2.44985134 -4.533433226
#>  [346,] -0.89839873 -2.766728884
#>  [347,]  1.24929277 -0.295848223
#>  [348,]  1.24929277  2.544514108
#>  [349,] -0.82166460 -3.640349901
#>  [350,]  0.46189956 -1.064478399
#>  [351,] -0.29840146 -3.593781147
#>  [352,] -0.13809771 -3.190687166
#>  [353,] -1.16628719  0.001967399
#>  [354,]  1.24929277 -2.110858554
#>  [355,]  1.24929277 -1.791715853
#>  [356,]  1.24929277 -2.169894431
#>  [357,] -1.84985407 -2.388437026
#>  [358,]  1.24929277  1.275614003
#>  [359,]  0.46189956  2.268905012
#>  [360,] -1.68579195  0.452074942
#>  [361,] -1.00222507  2.355195656
#>  [362,]  0.38516542  0.355936171
#>  [363,]  0.64929550 -0.059572714
#>  [364,] -0.13809771 -3.206111034
#>  [365,] -2.55367767 -4.533433226
#>  [366,]  0.46189956  0.368632741
#>  [367,] -1.58572399  0.475083159
#>  [368,] -1.68579195 -1.980604799
#>  [369,]  1.24929277 -3.748656279
#>  [370,] -0.30215984 -1.765498870
#>  [371,] -0.89839873 -2.330969008
#>  [372,]  1.24929277 -2.266285083
#>  [373,] -2.63724729 -3.779975038
#>  [374,]  1.24929277  1.965759836
#>  [375,] -1.76628446 -4.533433226
#>  [376,] -1.58572399 -4.533433226
#>  [377,]  0.64929550 -3.024293651
#>  [378,] -0.21483185 -1.647637587
#>  [379,]  1.24929277  2.596518200
#>  [380,]  0.38516542 -1.182339682
#>  [381,] -2.55367767 -1.973708496
#>  [382,] -0.21483185 -1.117086078
#>  [383,] -3.23724456 -4.533433226
#>  [384,] -3.23724456 -1.647637587
#>  [385,]  1.24929277 -0.113855580
#>  [386,]  0.56572589 -2.089094701
#>  [387,] -0.13809771  3.619205493
#>  [388,] -2.44985134 -2.898241145
#>  [389,] -1.68579195 -2.864316742
#>  [390,] -2.44985134 -3.598234276
#>  [391,] -1.68955032 -3.640349901
#>  [392,]  0.38516542 -2.908907447
#>  [393,]  0.56572589 -1.947239633
#>  [394,]  0.64929550  1.808485822
#>  [395,] -1.08579468 -3.748656279
#>  [396,] -0.03427138  1.318132947
#>  [397,] -3.23724456 -3.944145355
#>  [398,] -3.23724456 -2.139128238
#>  [399,]  1.24929277  1.729567397
#>  [400,] -1.95368040 -2.336292868
#>  [401,]  0.56572589 -0.269931926
#>  [402,]  0.46189956 -1.365895687
#>  [403,] -2.63724729 -1.674106450
#>  [404,]  0.56572589 -2.700697821
#>  [405,] -2.37311720 -3.791324470
#>  [406,] -1.95368040 -2.898241145
#>  [407,] -0.03427138 -2.855572954
#>  [408,]  0.46189956 -2.297603841
#>  [409,]  1.24929277 -2.169894431
#>  [410,] -1.58572399 -3.024293651
#>  [411,] -3.23724456 -1.200029100
#>  [412,] -1.68579195 -4.533433226
#>  [413,]  1.24929277  2.679553413
#>  [414,]  1.24929277 -1.006938906
#>  [415,]  0.64929550 -3.480372992
#>  [416,] -1.08579468 -1.920374003
#>  [417,]  1.24929277 -1.652090716
#>  [418,] -2.44985134 -4.533433226
#>  [419,] -3.23724456 -1.985057928
#>  [420,] -3.23724456 -3.117093092
#>  [421,] -0.13809771 -3.480372992
#>  [422,] -1.84985407 -2.891085121
#>  [423,] -3.23724456 -2.813457328
#>  [424,] -3.23724456 -0.980721923
#>  [425,] -2.44985134 -4.533433226
#>  [426,]  1.24929277 -0.516201814
#>  [427,] -2.44985134 -3.791324470
#>  [428,] -1.77311993 -2.118091198
#>  [429,] -2.63724729 -2.079539795
#>  [430,] -1.68579195 -2.881236994
#>  [431,] -3.23724456 -2.727166684
#>  [432,] -1.68579195 -3.779975038
#>  [433,]  1.24929277  2.040518875
#>  [434,] -1.84985407 -2.079539795
#>  [435,] -2.55367767 -4.533433226
#>  [436,] -2.63724729 -3.163918788
#>  [437,] -1.00222507 -2.864316742
#>  [438,] -1.76628446 -2.924331314
#>  [439,] -2.37311720 -3.640349901
#>  [440,] -0.82166460 -2.009917522
#>  [441,] -0.40222780 -4.533433226
#>  [442,] -2.44985134 -4.533433226
#>  [443,] -1.08579468 -2.147319461
#>  [444,] -1.68579195 -3.791324470
#>  [445,]  0.64929550 -2.144782956
#>  [446,] -3.23724456 -3.917376977
#>  [447,] -1.08955305 -1.981951937
#>  [448,] -3.23724456 -3.008946404
#>  [449,] -2.37311720 -0.477265491
#>  [450,]  0.64929550  0.410051507
#>  [451,] -3.23724456 -2.886891712
#>  [452,] -2.37311720 -1.978598764
#>  [453,] -0.22166733 -1.971233417
#>  [454,] -2.55367767 -3.748656279
#>  [455,]  0.56572589 -1.362404890
#>  [456,] -0.89839873 -2.945368354
#>  [457,]  0.64929550  2.177452709
#>  [458,]  0.64929550  4.986748754
#>  [459,] -0.03427138 -0.417491344
#>  [460,] -2.63724729 -3.944145355
#>  [461,]  1.24929277  3.632658183
#>  [462,] -2.44985134 -1.480244107
#>  [463,] -0.89839873 -3.640349901
#>  [464,] -0.89839873 -3.706380963
#>  [465,] -0.03427138 -1.339829920
#>  [466,]  1.24929277 -0.535352627
#>  [467,]  1.24929277  2.592869894
#>  [468,]  1.24929277  0.016412572
#>  [469,] -2.55367767 -3.008946404
#>  [470,]  0.56572589 -2.726914804
#>  [471,]  0.64929550 -1.124640499
#>  [472,] -3.23724456 -3.024293651
#>  [473,] -1.68579195 -2.574630917
#>  [474,] -2.37311720 -1.594757769
#>  [475,] -1.00222507 -3.917376977
#>  [476,] -0.90215711 -4.533433226
#>  [477,] -0.13809771 -3.010621958
#>  [478,] -2.55367767 -3.917376977
#>  [479,]  1.24929277  2.917317805
#>  [480,]  0.64929550 -3.051062030
#>  [481,] -1.76628446 -2.891085121
#>  [482,] -3.23724456 -2.071348572
#>  [483,]  1.24929277 -1.171206474
#>  [484,]  0.56572589 -1.559569520
#>  [485,]  0.56572589  1.030224544
#>  [486,]  0.56572589 -0.481782394
#>  [487,]  1.24929277 -1.331086132
#>  [488,] -0.21483185  1.965038083
#>  [489,] -3.23724456 -3.791324470
#>  [490,] -1.08579468 -2.923352620
#>  [491,]  0.64929550 -2.769282310
#>  [492,]  0.46189956 -1.920374003
#>  [493,]  1.24929277  0.517379190
#>  [494,]  1.24929277 -1.507269427
#>  [495,] -2.55367767 -2.139128238
#>  [496,] -0.98572672 -2.042525946
#>  [497,]  0.46189956 -1.393861272
#>  [498,]  0.46189956 -1.074718571
#>  [499,] -3.23724456 -1.760237404
#>  [500,] -3.23724456 -0.395984432
#>  [501,]  0.64929550 -3.944145355
#>  [502,]  0.56572589 -0.433250162
#>  [503,]  0.56572589 -0.555052973
#>  [504,] -0.30215984  2.413861372
#>  [505,]  0.64929550  0.101453792
#>  [506,]  1.24929277 -1.477148366
#>  [507,] -0.21483185  3.814301694
#>  [508,]  1.24929277  4.446257756
#>  [509,]  1.24929277  2.960481053
#>  [510,] -0.13809771 -2.995198091
#>  [511,]  0.56572589  1.133435440
#>  [512,]  1.24929277  3.783458881
#>  [513,] -0.03427138  3.113204225
#>  [514,]  1.24929277 -1.951692762
#>  [515,]  0.64929550 -2.059999140
#>  [516,] -0.13809771 -1.373754323
#>  [517,]  1.24929277  0.085156751
#>  [518,] -3.23724456 -2.010169403
#>  [519,] -3.23724456 -4.533433226
#>  [520,] -1.95368040 -2.844776087
#>  [521,]  1.24929277 -1.178265247
#>  [522,] -0.90215711 -2.096460047
#>  [523,] -2.44985134 -3.662365635
#>  [524,] -2.55367767 -3.202036599
#>  [525,]  1.24929277 -0.588977376
#>  [526,] -1.68579195 -1.955886170
#>  [527,] -2.55367767 -0.935093434
#>  [528,]  1.24929277 -0.577627943
#>  [529,] -3.23724456  0.306711594
#>  [530,] -0.21483185  3.490246658
#>  [531,]  1.24929277  0.276766356
#>  [532,] -1.00222507 -1.362542514
#>  [533,] -2.63724729 -0.542248931
#>  [534,] -1.68955032 -2.059999140
#>  [535,]  1.24929277  3.570408619
#>  [536,] -0.21483185  1.255741928
#>  [537,]  1.24929277  5.188366512
#>  [538,] -1.00222507 -2.374984336
#>  [539,]  1.24929277  1.833597297
#>  [540,]  1.24929277  4.248714433
#>  [541,]  1.24929277  0.150713849
#>  [542,]  1.24929277  3.442699209
#>  [543,] -1.68955032 -1.317890384
#>  [544,]  0.64929550  2.044244178
#>  [545,] -2.55367767  1.491682486
#>  [546,]  1.24929277 -2.855572954
#>  [547,] -0.13809771 -4.533433226
#>  [548,]  1.24929277 -1.058050200
#>  [549,]  0.38516542  0.959577516
#>  [550,]  1.24929277 -4.533433226
#>  [551,] -0.30215984  1.616387855
#>  [552,] -1.16628719 -2.840322958
#>  [553,] -0.89839873 -1.947239633
#>  [554,]  1.24929277 -2.297603841
#>  [555,] -0.21483185 -1.124380778
#>  [556,] -0.03427138 -2.705150950
#>  [557,] -0.21483185 -1.093495612
#>  [558,] -0.21483185 -3.917376977
#>  [559,] -0.21483185 -4.533433226
#>  [560,] -0.89839873 -0.282576278
#>  [561,]  0.56572589 -2.968346642
#>  [562,] -0.30215984  0.560852204
#>  [563,]  1.24929277 -3.593781147
#>  [564,] -0.30215984 -1.245666219
#>  [565,] -3.23724456 -2.952922775
#>  [566,]  1.24929277  0.453576247
#>  [567,]  1.24929277 -1.203550593
#>  [568,] -0.30215984  0.866827707
#>  [569,] -3.23724456 -2.722713555
#>  [570,]  1.24929277 -1.132258064
#>  [571,] -0.30215984  0.560852204
#>  [572,]  1.24929277  3.045424560
#>  [573,] -0.21483185 -0.485430700
#>  [574,]  1.24929277  1.705803060
#>  [575,]  1.24929277  0.956064652
#>  [576,]  0.38516542  0.560852204
#>  [577,]  1.24929277  0.560852204
#>  [578,] -1.58572399 -3.480372992
#>  [579,] -0.13809771 -3.662365635
#>  [580,]  0.64929550  1.181080140
#>  [581,]  1.24929277  1.346499882
#>  [582,]  1.24929277 -2.877588688
#>  [583,]  1.24929277 -0.694518419
#>  [584,]  1.24929277  3.845620452
#>  [585,]  1.24929277  3.697168237
#>  [586,] -1.84985407 -1.239790963
#>  [587,]  1.24929277 -3.779975038
#>  [588,]  0.64929550 -0.825038453
#>  [589,] -0.03427138 -3.640349901
#>  [590,]  0.64929550 -2.705150950
#>  [591,] -0.03427138 -1.963042194
#>  [592,]  1.24929277  0.912087144
#>  [593,] -0.90215711 -2.813457328
#>  [594,]  0.56572589  1.339360780
#>  [595,]  1.24929277  4.595004206
#>  [596,] -3.23724456 -4.533433226
#>  [597,]  1.24929277  0.983089972
#>  [598,] -1.08579468 -2.726914804
#>  [599,] -0.21483185  0.394053981
#>  [600,]  1.24929277 -0.459766660
#>  [601,] -0.30215984  0.502200669
#>  [602,] -0.89839873 -3.480372992
#>  [603,] -0.21483185 -0.160513746
#>  [604,]  1.24929277  0.674981621
#>  [605,] -1.68579195 -3.779975038
#>  [606,]  1.24929277  4.556886395
#>  [607,] -1.00222507 -2.726914804
#>  [608,] -0.21483185  1.657511253
#>  [609,]  1.24929277 -0.385068247
#>  [610,] -1.68955032 -2.989383682
#>  [611,] -0.89839873 -1.352942175
#>  [612,]  1.24929277  2.326199159
#>  [613,]  0.38516542  2.154740115
#>  [614,]  1.24929277 -1.716222076
#>  [615,] -2.44985134 -3.190687166
#>  [616,]  0.64929550 -3.662365635
#>  [617,] -3.23724456 -3.706380963
#>  [618,] -1.68579195  0.008206951
#>  [619,] -1.68579195  0.008206951
#>  [620,]  0.64929550 -2.089094701
#>  [621,]  1.24929277  3.527642716
#>  [622,] -0.21483185 -4.533433226
#>  [623,]  0.64929550 -2.891085121
#>  [624,]  0.46189956  3.600301640
#>  [625,]  1.24929277  1.373268261
#>  [626,]  0.64929550  1.976857388
#>  [627,]  0.64929550 -0.359174760
#>  [628,] -0.40222780 -0.203650980
#>  [629,]  1.24929277  1.696864091
#>  [630,]  1.24929277  0.588491313
#>  [631,]  1.24929277  1.666540376
#>  [632,] -0.13809771  0.023004439
#>  [633,] -3.23724456 -3.706380963
#>  [634,] -2.37311720 -3.748656279
#>  [635,]  0.46189956  4.434908324
#>  [636,]  0.56572589 -3.748656279
#>  [637,] -0.21483185  0.191249191
#>  [638,] -1.00222507 -3.006547523
#>  [639,]  1.24929277  5.188366512
#>  [640,]  1.24929277  2.870575179
#>  [641,]  1.24929277 -1.470551578
#>  [642,] -3.23724456 -1.791715853
#>  [643,]  1.24929277  2.910337042
#>  [644,] -0.89839873  3.882081359
#>  [645,] -2.37311720 -3.090324714
#>  [646,]  0.64929550 -1.892308119
#>  [647,] -0.30215984  1.961306707
#>  [648,]  1.24929277  0.206756129
#>  [649,]  1.24929277  4.228280979
#>  [650,] -1.58572399 -4.533433226
#>  [651,]  1.24929277  3.527642716
#>  [652,] -1.95368040 -2.769282310
#>  [653,] -0.22166733 -3.917376977
#>  [654,] -1.68579195 -1.217775229
#>  [655,]  1.24929277 -0.645407609
#>  [656,]  1.24929277  0.304258582
#>  [657,] -0.89839873 -2.609305401
#>  [658,] -3.23724456 -0.862860640
#>  [659,] -0.13809771 -4.533433226
#>  [660,] -0.03427138 -2.239516704
#>  [661,]  0.46189956 -2.705150950
#>  [662,]  0.64929550  4.446257756
#>  [663,] -0.21483185  2.045271519
#>  [664,]  1.24929277  3.661480809
#>  [665,]  1.24929277  3.830201507
#>  [666,] -2.44985134 -2.977724897
#>  [667,] -3.23724456 -4.533433226
#>  [668,] -3.23724456 -3.593781147
#>  [669,]  1.24929277  2.132081652
#>  [670,]  0.64929550  0.757212011
#>  [671,] -0.21483185  4.471369231
#>  [672,] -0.89839873 -0.954862703
#>  [673,] -0.13809771  1.058123374
#>  [674,] -0.29840146 -0.562208997
#>  [675,]  0.56572589 -2.886891712
#>  [676,] -2.55367767 -3.816435945
#>  [677,] -3.23724456 -1.100165826
#>  [678,] -2.37311720 -2.587289667
#>  [679,]  0.64929550 -2.844776087
#>  [680,] -1.68955032 -2.920256879
#>  [681,]  1.24929277  5.188366512
#>  [682,] -2.44985134 -0.898632527
#>  [683,] -1.76628446 -2.102114765
#>  [684,]  0.64929550 -1.971233417
#>  [685,]  1.24929277  0.909495898
#>  [686,]  0.46189956 -1.239790963
#>  [687,]  0.56572589  3.113204225
#>  [688,]  0.38516542  1.155066591
#>  [689,] -0.13809771 -2.968346642
#>  [690,] -0.89839873 -3.480372992
#>  [691,] -2.63724729 -2.855572954
#>  [692,]  0.64929550  4.248714433
#>  [693,] -1.00222507  0.447065356
#>  [694,] -0.21483185  0.502200669
#>  [695,] -3.23724456 -3.640349901
#>  [696,]  1.24929277  1.630833028
#>  [697,] -1.58572399 -0.784485102
#>  [698,] -0.03427138 -1.076172042
#>  [699,]  0.38516542  0.657075802
#>  [700,] -0.21483185  3.632658183
#>  [701,]  0.64929550  0.156289590
#>  [702,]  0.38516542  5.926400834
#>  [703,]  1.24929277  1.528143393
#>  [704,]  0.46189956  0.635060068
#>  [705,]  0.64929550 -3.944145355
#>  [706,] -0.03427138 -3.593781147
#>  [707,] -1.08955305 -3.004493275
#>  [708,]  0.56572589  2.130883954
#>  [709,] -0.98572672  1.541596083
#>  [710,] -0.98572672 -1.747578654
#>  [711,]  1.24929277  0.871280836
#>  [712,]  1.24929277  1.023708875
#>  [713,] -1.68955032 -3.791324470
#>  [714,] -1.76628446 -4.533433226
#>  [715,] -2.37311720 -2.079539795
#>  [716,] -0.03427138 -3.480372992
#>  [717,] -0.21483185 -0.113855580
#>  [718,]  1.24929277  2.761590592
#>  [719,]  1.24929277 -1.986405065
#>  [720,]  1.24929277  1.064371030
#>  [721,] -0.29840146 -0.346707637
#>  [722,]  0.38516542 -3.944145355
#>  [723,]  0.56572589 -2.255488216
#>  [724,]  1.24929277  3.729260475
#>  [725,] -0.29840146  2.355195656
#>  [726,]  0.38516542  2.245140675
#>  [727,]  0.38516542 -3.046309386
#>  [728,] -0.13809771 -2.037264479
#>  [729,] -0.89839873 -2.110858554
#>  [730,]  0.64929550 -3.748656279
#>  [731,] -0.21483185 -1.163015251
#>  [732,] -0.21483185  0.475432291
#>  [733,] -0.89839873 -2.813457328
#>  [734,]  1.24929277  2.266177714
#>  [735,]  1.24929277  5.926400834
#>  [736,]  0.38516542  2.200320524
#>  [737,]  1.24929277 -3.748656279
#>  [738,]  1.24929277 -1.233199096
#>  [739,] -0.21483185 -2.908907447
#>  [740,]  1.24929277 -0.023111807
#>  [741,]  1.24929277  3.045424560
#>  [742,]  0.64929550 -3.328089105
#>  [743,] -0.03427138 -4.533433226
#>  [744,]  1.24929277  0.691180949
#>  [745,]  0.46189956 -2.278200809
#>  [746,] -2.37311720 -4.533433226
#>  [747,]  0.64929550  1.346499882
#>  [748,]  0.64929550 -3.917376977
#>  [749,] -0.89839873 -4.533433226
#>  [750,] -1.68955032 -4.533433226
#>  [751,] -2.37311720 -4.533433226
#>  [752,] -0.40222780 -2.908907447
#>  [753,]  0.64929550  0.372038247
#>  [754,] -2.44985134 -3.662365635
#>  [755,] -1.95368040 -4.533433226
#>  [756,] -0.03427138  1.434647093
#>  [757,]  1.24929277  1.333706061
#>  [758,] -1.68955032 -0.999224794
#>  [759,]  0.46189956 -0.264670459
#>  [760,] -0.29840146 -2.015824121
#>  [761,]  1.24929277  3.661480809
#>  [762,]  1.24929277  4.510060699
#>  [763,]  0.38516542 -1.046752986
#>  [764,] -1.58572399 -0.712438636
#>  [765,] -0.82166460 -1.080625170
#>  [766,] -1.84985407 -2.545174041
#>  [767,]  1.24929277  2.757167025
#>  [768,]  0.38516542 -1.552089578
#>  [769,]  0.56572589  5.926400834
#>  [770,]  1.24929277  0.978080386
#>  [771,]  1.24929277  3.531717151
#>  [772,] -0.13809771 -0.304592012
#>  [773,] -1.00222507 -0.224729566
#>  [774,]  1.24929277  1.918540804
#>  [775,]  0.56572589 -2.148857391
#>  [776,] -0.89839873 -2.102667331
#>  [777,] -1.58572399 -3.004493275
#>  [778,]  1.24929277  1.915813506
#>  [779,] -1.76628446 -3.598234276
#>  [780,] -1.00222507  1.415543576
#>  [781,]  0.64929550 -1.951692762
#>  [782,] -0.03427138  1.388775198
#>  [783,] -0.03427138 -0.324768523
#>  [784,]  0.64929550 -1.213658440
#>  [785,]  0.56572589 -1.107531172
#>  [786,] -2.55367767 -4.533433226
#>  [787,] -2.37311720 -1.022203083
#>  [788,]  0.64929550 -0.999224794
#>  [789,]  0.56572589 -1.517453894
#>  [790,]  1.24929277  3.655352126
#>  [791,] -0.90215711 -0.201404514
#>  [792,]  1.24929277  3.745258000
#>  [793,]  1.24929277  3.830201507
#>  [794,]  1.24929277  4.568235828
#>  [795,]  0.46189956  2.205675727
#>  [796,] -0.90215711 -2.291949123
#>  [797,]  0.56572589  1.192429573
#>  [798,]  1.24929277 -0.527798206
#>  [799,]  1.24929277  4.382351289
#>  [800,] -2.55367767 -1.561384763
#>  [801,]  0.64929550 -2.844776087
#>  [802,]  1.24929277  3.511283698
#>  [803,]  1.24929277 -0.245766606
#>  [804,]  0.46189956  3.619205493
#>  [805,] -0.03427138 -2.102667331
#>  [806,]  0.64929550  3.486172223
#>  [807,] -0.13809771 -1.802512720
#>  [808,]  0.56572589 -2.042525946
#>  [809,] -0.89839873 -3.779975038
#>  [810,]  0.64929550 -0.145863358
#>  [811,]  0.46189956 -1.993249151
#>  [812,]  1.24929277  0.553779010
#>  [813,] -0.03427138 -4.533433226
#>  [814,] -0.89839873 -3.598234276
#>  [815,]  1.24929277  3.531717151
#>  [816,]  1.24929277  5.926400834
#>  [817,]  0.64929550  1.153235816
#>  [818,] -0.03427138  0.278015781
#>  [819,]  1.24929277  5.926400834
#>  [820,]  1.24929277  2.679553413
#>  [821,]  0.64929550  3.856969885
#>  [822,]  1.24929277  5.310344584
#>  [823,] -0.21483185  3.648455823
#>  [824,] -0.21483185  2.714891709
#>  [825,]  0.64929550  0.410051507
#>  [826,]  0.64929550 -1.559569520
#>  [827,]  0.56572589  4.593347302
#>  [828,]  1.24929277  5.184292077
#>  [829,]  1.24929277  0.372038247
#>  [830,]  1.24929277  4.986748754
#>  [831,]  1.24929277  2.620114218
#>  [832,]  1.24929277  2.157652333
#>  [833,]  1.24929277  3.729260475
#>  [834,]  1.24929277  2.386514415
#>  [835,]  1.24929277  3.526393291
#>  [836,]  1.24929277  3.393197522
#>  [837,]  1.24929277  0.293439648
#>  [838,]  0.46189956  2.312920340
#>  [839,]  0.46189956  2.141110775
#>  [840,]  1.24929277  5.926400834
#>  [841,]  1.24929277  4.471369231
#>  [842,]  0.64929550 -3.944145355
#>  [843,]  1.24929277 -2.270835463
#>  [844,]  0.56572589  4.382351289
#>  [845,] -0.21483185 -3.640349901
#>  [846,]  1.24929277  3.003149244
#>  [847,] -1.68579195 -0.820585324
#>  [848,] -1.08579468 -1.335636512
#>  [849,] -0.21483185  2.746940204
#>  [850,] -0.40222780 -2.148976365
#>  [851,]  0.38516542  2.130883954
#>  [852,]  1.24929277 -1.213658440
#>  [853,]  1.24929277  2.571406726
#>  [854,]  0.64929550  0.674981621
#>  [855,] -1.08955305  0.517379190
#>  [856,]  1.24929277 -1.802512720
#>  [857,] -0.82166460 -0.295848223
#>  [858,]  0.46189956 -1.080625170
#>  [859,]  1.24929277  3.596227205
#>  [860,] -3.23724456 -3.706380963
#>  [861,]  1.24929277  2.132540858
#>  [862,]  1.24929277  1.079480624
#>  [863,]  1.24929277  2.564425962
#>  [864,]  1.24929277  1.305488595
#>  [865,]  0.56572589 -0.390722966
#>  [866,] -0.21483185 -0.113855580
#>  [867,] -0.13809771  1.563053363
#>  [868,]  0.64929550 -3.074327189
#>  [869,]  0.46189956 -0.114029452
#>  [870,]  0.46189956 -1.080625170
#>  [871,] -2.63724729 -1.107146589
#>  [872,]  0.56572589  2.523916354
#>  [873,] -1.08579468 -0.889329503
#>  [874,]  0.56572589  2.954352370
#>  [875,]  0.46189956  0.399517907
#>  [876,]  0.46189956 -0.353940282
#>  [877,]  0.38516542  0.612829308
#>  [878,] -0.22166733  0.606237442
#>  [879,]  0.46189956  0.409072812
#>  [880,]  0.46189956 -0.193963373
#>  [881,] -1.77311993  2.237355088
#>  [882,] -3.23724456 -2.763375711
#>  [883,] -2.55367767 -1.989132363
#>  [884,]  1.24929277  5.184292077
#>  [885,]  1.24929277  3.729260475
#>  [886,] -1.84985407 -1.393861272
#>  [887,] -0.21483185 -2.291949123
#>  [888,]  1.24929277 -0.527546325
#>  [889,]  1.24929277  2.991799811
#>  [890,]  0.64929550  2.083284779
#>  [891,]  1.24929277  3.506605676
#>  [892,] -0.03427138 -4.533433226
#>  [893,]  0.64929550 -1.791715853
#>  [894,] -3.23724456 -4.533433226
#>  [895,] -3.23724456 -2.251035087
#>  [896,] -3.23724456 -4.533433226
#>  [897,]  0.64929550 -0.811253480
#>  [898,]  0.64929550 -0.315388879
#>  [899,]  1.24929277  3.653695223
#>  [900,] -0.03427138 -3.090324714
#>  [901,] -0.21483185 -1.372097419
#>  [902,] -0.30215984 -0.108041171
#>  [903,]  1.24929277  2.270252149
#>  [904,] -0.03427138  1.919887941
#>  [905,] -0.89839873 -3.037866282
#>  [906,]  1.24929277  0.475432291
#>  [907,] -1.76628446  2.726122168
#>  [908,]  0.64929550 -0.781023125
#>  [909,]  0.64929550 -2.166798690
#>  [910,]  1.24929277  2.811450258
#>  [911,]  1.24929277  2.355195656
#>  [912,] -0.03427138 -1.470711268
#>  [913,]  1.24929277  3.139972603
#>  [914,]  0.46189956 -2.813297638
#>  [915,] -1.16628719 -2.855572954
#>  [916,]  1.24929277  3.195654198
#>  [917,]  0.64929550  1.192429573
#>  [918,]  1.24929277 -0.221449326
#>  [919,] -0.89839873 -3.132600030
#>  [920,] -0.89839873 -1.978598764
#>  [921,] -1.68579195 -4.533433226
#>  [922,]  0.56572589  2.044593310
#>  [923,] -0.89839873  1.415543576
#>  [924,] -1.95368040  0.544475985
#>  [925,] -0.98572672  2.044593310
#>  [926,] -3.23724456 -2.738264236
#>  [927,] -1.68579195 -1.447976608
#>  [928,] -0.30215984 -1.652090716
#>  [929,] -0.03427138 -0.305143408
#>  [930,]  1.24929277  3.661480809
#>  [931,]  1.24929277 -0.113855580
#>  [932,]  1.24929277  2.608420575
#>  [933,] -0.29840146  1.302484554
#>  [934,]  1.24929277 -0.010767331
#>  [935,]  0.38516542 -2.075423007
#>  [936,] -0.89839873  0.434421004
#>  [937,]  0.64929550  1.695077668
#>  [938,]  1.24929277  5.141623887
#>  [939,]  1.24929277 -1.426536250
#>  [940,]  0.64929550 -1.428435953
#>  [941,]  0.56572589 -0.820585324
#>  [942,]  0.38516542  0.591044740
#>  [943,]  1.24929277  1.441852749
#>  [944,] -0.89839873 -1.652090716
#>  [945,] -1.00222507 -2.110858554
#>  [946,]  0.46189956 -1.217775229
#>  [947,]  1.24929277  2.260455248
#>  [948,]  0.56572589  3.089939066
#>  [949,]  0.64929550 -1.335636512
#>  [950,]  1.24929277  5.337112962
#>  [951,]  1.24929277  0.321012849
#>  [952,]  1.24929277  1.528143393
#>  [953,]  1.24929277  4.446257756
#>  [954,] -0.13809771  0.273346427
#>  [955,] -2.63724729 -2.855572954
#>  [956,] -2.37311720 -2.726914804
#>  [957,] -3.23724456 -3.163918788
#>  [958,]  1.24929277  2.679553413
#>  [959,]  1.24929277 -0.847054187
#>  [960,] -1.68579195 -1.855847212
#>  [961,]  1.24929277  3.637111312
#>  [962,]  1.24929277  4.620115681
#>  [963,]  1.24929277  3.856969885
#>  [964,] -0.13809771 -0.597602191
#>  [965,]  0.56572589  2.132081652
#>  [966,]  1.24929277  5.184292077
#>  [967,] -2.37311720 -4.533433226
#>  [968,] -2.44985134 -2.329312104
#>  [969,]  0.56572589 -1.317890384
#>  [970,]  1.24929277  0.393802101
#>  [971,] -3.23724456 -4.533433226
#>  [972,]  0.38516542 -0.280254017
#>  [973,] -0.03427138 -3.010621958
#>  [974,]  1.24929277 -1.342532815
#>  [975,]  0.64929550 -0.939438835
#>  [976,]  0.46189956 -0.852063773
#>  [977,]  1.24929277  5.310344584
#>  [978,] -0.22166733 -2.255488216
#>  [979,] -3.23724456 -4.533433226
#>  [980,]  1.24929277 -0.550859565
#>  [981,]  1.24929277  2.117431265
#>  [982,]  1.24929277  2.874190409
#>  [983,] -0.03427138 -1.071070265
#>  [984,]  0.46189956 -1.244543607
#>  [985,]  0.64929550 -0.291395094
#>  [986,]  0.64929550 -0.582170484
#>  [987,]  1.24929277  2.117431265
#>  [988,]  0.64929550  0.492267716
#>  [989,]  1.24929277 -1.166915815
#>  [990,]  1.24929277  3.065212174
#>  [991,] -3.23724456 -4.533433226
#>  [992,]  0.56572589 -2.270835463
#>  [993,]  1.24929277 -0.962763887
#>  [994,]  0.46189956 -2.844776087
#>  [995,] -0.21483185  1.415543576
#>  [996,] -1.68955032 -3.779975038
#>  [997,] -0.89839873 -3.598234276
#>  [998,] -1.16628719  0.394053981
#>  [999,]  1.24929277  4.599078641
#> [1000,] -0.82166460 -3.640349901
#> [1001,] -1.08579468 -3.640349901
#> [1002,]  1.24929277  1.415543576
#> [1003,]  1.24929277  1.373268261
#> [1004,]  1.24929277  3.680463601
#> [1005,] -1.68579195 -1.337431039
#> [1006,] -3.23724456 -3.816435945
#> [1007,] -0.13809771  2.355195656
#> [1008,]  0.56572589  3.029917622
#> [1009,]  0.38516542 -1.116834197
#> [1010,] -0.21483185 -3.779975038
#> [1011,]  0.56572589 -0.220655131
#> [1012,]  0.46189956 -1.182339682
#> [1013,] -0.21483185  1.879352599
#> [1014,]  1.24929277  3.576537302
#> [1015,]  1.24929277  3.576537302
#> [1016,] -0.29840146 -1.230242351
#> [1017,]  1.24929277  5.337112962
#> [1018,]  1.24929277 -3.662365635
#> [1019,]  1.24929277  1.770270760
#> [1020,]  1.24929277  2.944483528
#> [1021,] -2.55367767 -2.855572954
#> [1022,]  0.46189956 -4.533433226
#> [1023,]  1.24929277  5.184292077
#> [1024,] -0.21483185 -0.417651034
#> [1025,] -0.21483185  2.523916354
#> [1026,]  0.46189956  0.899390574
#> [1027,] -3.23724456 -3.090324714
#> [1028,] -0.21483185  1.813321537
#> [1029,]  0.56572589  0.121806733
#> [1030,]  1.24929277  3.655352126
#> [1031,] -2.37311720 -2.025341389
#> [1032,]  1.24929277 -0.113855580
#> [1033,]  1.24929277  4.721056713
#> [1034,] -1.58572399 -2.982178026
#> [1035,]  1.24929277  3.729260475
#> [1036,]  1.24929277 -0.027564936
#> [1037,]  0.46189956 -1.920374003
#> [1038,]  0.64929550 -4.533433226
#> [1039,]  0.64929550  2.326373030
#> [1040,]  0.56572589  0.412105755
#> [1041,]  0.56572589  1.180332611
#> [1042,] -0.03427138 -1.228493748
#> [1043,]  1.24929277  1.278720217
#> [1044,]  1.24929277  3.225489768
#> [1045,] -1.08955305  0.674729740
#> [1046,]  0.64929550  1.222453382
#> [1047,]  0.64929550  2.975802286
#> [1048,]  1.24929277  4.620115681
#> [1049,] -0.21483185 -2.881236994
#> [1050,]  1.24929277  3.006939333
#> [1051,] -3.23724456 -2.059999140
#> [1052,] -2.37311720 -2.106308174
#> [1053,] -3.23724456 -2.160591407
#> [1054,]  0.46189956  3.576537302
#> [1055,] -0.29840146 -0.920648262
#> [1056,]  1.24929277  2.792812099
#> [1057,]  1.24929277  4.593347302
#> [1058,]  0.56572589  2.172302720
#> [1059,] -0.82166460 -3.640349901
#> [1060,]  0.64929550 -0.253572907
#> [1061,] -0.30215984  2.216696742
#> [1062,]  0.38516542  1.962556132
#> [1063,]  0.64929550 -0.522793681
#> [1064,]  0.38516542  3.685818804
#> [1065,]  0.56572589 -3.200379695
#> [1066,] -0.21483185 -2.891085121
#> [1067,]  1.24929277 -1.335636512
#> [1068,]  1.24929277 -0.432998282
#> [1069,]  1.24929277  4.620115681
#> [1070,] -3.23724456 -2.160591407
#> [1071,]  0.56572589  0.412105755
#> [1072,] -2.63724729 -1.920374003
#> [1073,]  1.24929277  5.184292077
#> [1074,]  0.56572589  0.502200669
#> [1075,]  1.24929277 -0.863413205
#> [1076,]  1.24929277  4.467294796
#> [1077,] -0.03427138 -3.779975038
#> [1078,]  0.64929550 -3.006547523
#> [1079,]  0.64929550  2.964271142
#> [1080,]  1.24929277  1.696864091
#> [1081,] -0.03427138 -0.244018002
#> [1082,]  1.24929277  2.220771178
#> [1083,] -0.13809771 -3.593781147
#> [1084,]  0.64929550 -2.182222558
#> [1085,] -0.03427138 -1.377828758
#> [1086,]  0.56572589  3.139972603
#> [1087,]  1.24929277  3.543640241
#> [1088,]  0.38516542  3.055029096
#> [1089,]  0.56572589 -1.006779215
#> [1090,]  1.24929277  1.441852749
#> [1091,]  1.24929277  3.628583748
#> [1092,]  1.24929277  0.757212011
#> [1093,] -0.29840146  0.514845021
#> [1094,]  0.64929550 -2.840322958
#> [1095,] -0.21483185 -1.372097419
#> [1096,]  1.24929277 -1.175659603
#> [1097,] -1.84985407 -1.855847212
#> [1098,]  0.38516542  2.245140675
#> [1099,] -0.21483185 -3.480372992
#> [1100,]  1.24929277  2.003582023
#> [1101,] -0.21483185  2.046250214
#> [1102,] -3.23724456 -3.791324470
#> [1103,] -2.44985134 -2.059999140
#> [1104,] -0.21483185  0.868071815
#> [1105,] -1.77311993 -3.791324470
#> [1106,] -1.68955032 -2.247274926
#> [1107,] -0.89839873 -2.059999140
#> [1108,]  0.46189956  2.237355088
#> [1109,] -3.23724456 -4.533433226
#> [1110,]  1.24929277  3.661480809
#> [1111,] -1.68579195 -1.833831478
#> [1112,] -1.68579195 -4.533433226
#> [1113,]  1.24929277 -0.282576278
#> [1114,] -0.89839873 -1.984505363
#> [1115,]  1.24929277  0.398170770
#> [1116,]  1.24929277  5.926400834
#> [1117,]  1.24929277  4.620115681
#> [1118,]  1.24929277  1.233091728
#> [1119,]  0.38516542 -0.424807058
#> [1120,]  0.56572589  2.355195656
#> [1121,]  1.24929277  0.898398346
#> [1122,]  1.24929277  3.882081359
#> [1123,]  1.24929277  3.442699209
#> [1124,]  0.38516542  3.793063418
#> [1125,] -0.40222780  5.209403552
#> [1126,]  0.38516542  1.157752720
#> [1127,] -0.22166733  0.561722935
#> [1128,]  0.64929550 -0.391434052
#> [1129,]  0.38516542  2.632758570
#> [1130,]  1.24929277 -0.532953746
#> [1131,] -0.22166733  1.327687852
#> [1132,] -3.23724456 -1.546192061
#> [1133,]  0.64929550 -0.628487357
#> [1134,] -3.23724456 -3.795398905
#> [1135,] -0.89839873 -1.802512720
#> [1136,]  1.24929277  5.926400834
#> [1137,] -3.23724456 -1.195827852
#> [1138,] -0.22166733  3.266025110
#> [1139,]  1.24929277  3.531717151
#> [1140,]  1.24929277  5.926400834
#> [1141,] -0.82166460 -4.533433226
#> [1142,]  1.24929277  0.846056081
#> [1143,]  0.56572589  1.209826569
#> [1144,] -0.03427138 -3.090324714
#> [1145,] -1.68955032 -3.779975038
#> [1146,] -1.68955032 -3.779975038
#> [1147,] -0.03427138 -1.849255346
#> [1148,] -0.98572672 -2.363634903
#> [1149,] -2.44985134 -0.741895512
#> [1150,] -3.23724456 -2.017723824
#> [1151,] -0.21483185  1.196630821
#> [1152,]  0.64929550  1.104592098
#> [1153,]  1.24929277  3.628583748
#> [1154,] -0.03427138 -1.849255346
#> [1155,] -1.68955032 -3.779975038
#> [1156,] -0.03427138 -3.090324714
#> [1157,] -3.23724456 -2.017723824
#> [1158,] -0.21483185  1.196630821
#> [1159,] -0.82166460 -1.782253138
#> [1160,] -0.13809771  1.104592098
#> [1161,] -0.21483185 -1.372097419
#> [1162,] -1.68955032 -3.779975038
#> [1163,]  0.56572589  2.769174942
#> [1164,]  1.24929277  2.985902294
#> [1165,] -0.13809771  0.588491313
#> [1166,]  0.56572589  1.372416180
#> [1167,]  1.24929277 -0.920648262
#> [1168,]  0.64929550 -1.942137856
#> [1169,]  0.64929550 -0.113855580
#> [1170,]  1.24929277  2.086912369
#> [1171,] -0.21483185  1.307237198
#> [1172,]  0.38516542  0.434421004
#> [1173,]  0.64929550  1.537521648
#> [1174,] -1.68579195 -1.798612156
#> [1175,]  1.24929277  3.684916730
#> [1176,]  0.38516542  1.891423293
#> [1177,]  1.24929277 -2.851672390
#> [1178,] -0.21483185 -0.160513746
#> [1179,] -0.22166733  2.865747305
#> [1180,] -0.82166460  1.660403184
#> [1181,] -0.82166460  1.660403184
#> [1182,] -3.23724456 -3.598234276
#> [1183,]  0.56572589 -4.533433226
#> [1184,]  0.38516542  1.329252932
#> [1185,] -0.21483185  1.265121573
#> 
#> $scores$rot.scores
#> NULL
#> 
#> 
#> $loadings
#> Loading Method: Revised
#> 
#>       1     2    
#> wmt2  0.608 0.145
#> wmt1  0.402 0.106
#> wmt3  0.344 0.204
#> wmt5  0.318 0.234
#> wmt4  0.298 0.231
#> wmt9        0.523
#> wmt7        0.461
#> wmt15       0.441
#> wmt14       0.433
#> wmt6  0.177 0.429
#> wmt16       0.392
#> wmt8        0.391
#> wmt10  0.21 0.368
#> wmt12       0.334
#> wmt18       0.326
#> wmt13       0.304
#> wmt17 0.122 0.299
#> wmt11       0.277
#> Standardized loadings >= |0.10| are displayed. To change this 'minimum', use `print(net.loads_object, minimum = 0.10)`