Total Entropy Fit Index using Von Neumman's entropy (Quantum Information Theory) for correlation matrices

Computes the fit (TEFI) of a dimensionality structure using Von Neumman's entropy when the input is a correlation matrix. Lower values suggest better fit of a structure to the data.

Usage

tefi(data, structure = NULL, verbose = TRUE)

Arguments

data

Matrix, data frame, or *EGA class object. Matrix or data frame can be raw data or a correlation matrix. All *EGA objects are accepted. hierEGA input will produced the Generalized TEFI (see genTEFI)

structure

Numeric or character vector (length = ncol(data)). Can be theoretical factors or the structure detected by EGA

verbose

Boolean (length = 1). Whether messages and (insignificant) warnings should be output. Defaults to TRUE to see all messages and warnings for every function call. Set to FALSE to ignore messages and warnings

Value

Returns a data frame with columns:

Non-hierarchical Structure

VN.Entropy.Fit

The Total Entropy Fit Index using Von Neumman's entropy

Total.Correlation

The total correlation of the dataset

Average.Entropy

The average entropy of the dataset

Hierarchical Structure

VN.Entropy.Fit

The Generalized Total Entropy Fit Index using Von Neumman's entropy

Lower.Order.VN

Lower order (only) Total Entropy Fit Index

Higher.Order.VN

Higher order (only) Total Entropy Fit Index

References

Initial formalization and simulation
Golino, H., Moulder, R. G., Shi, D., Christensen, A. P., Garrido, L. E., Nieto, M. D., Nesselroade, J., Sadana, R., Thiyagarajan, J. A., & Boker, S. M. (2020). Entropy fit indices: New fit measures for assessing the structure and dimensionality of multiple latent variables. Multivariate Behavioral Research.

Author

Hudson Golino <hfg9s at virginia.edu>, Alexander P. Christensen <alexpaulchristensen@gmail.com>, and Robert Moulder <rgm4fd@virginia.edu>

Examples

# Load data
wmt <- wmt2[,7:24]

# Estimate EGA model
ega.wmt <- EGA(
  data = wmt, model = "glasso",
  plot.EGA = FALSE # no plot for CRAN checks
)

# Compute entropy indices for empirical EGA
tefi(ega.wmt)
#>   VN.Entropy.Fit Total.Correlation Average.Entropy
#> 1      -11.17104         -2.695968       -14.98371

# User-defined structure (with `EGA` object)
tefi(ega.wmt, structure = c(rep(1, 5), rep(2, 5), rep(3, 8)))
#>   VN.Entropy.Fit Total.Correlation Average.Entropy
#> 1      -10.35118         -5.597834       -20.04691