Total Correlation

Computes the total correlation of a dataset

Usage

totalCor(data, base = 2.718282)

Arguments

data

Matrix or data frame. Should consist only of variables to be used in the analysis

base

Numeric (length = 1). Base to use for entropy. Defaults to exp(1) or 2.718282

Value

Returns a list containing:

Ind.Entropies

Individual entropies for each variable

Joint.Entropy

The joint entropy of the dataset

Total.Cor

The total correlation of the dataset

Normalized

Total correlation divided by the sum of the individual entropies minus the maximum of the individual entropies

References

Formalization of total correlation
Watanabe, S. (1960). Information theoretical analysis of multivariate correlation. IBM Journal of Research and Development 4, 66-82.

Applied implementation
Felix, L. M., Mansur-Alves, M., Teles, M., Jamison, L., & Golino, H. (2021). Longitudinal impact and effects of booster sessions in a cognitive training program for healthy older adults. Archives of Gerontology and Geriatrics, 94, 104337.

Author

Hudson F. Golino <hfg9s at virginia.edu>

Examples

# Compute total correlation
totalCor(wmt2[,7:24])
#> $Ind.Entropies
#>  [1] 0.6248171 0.4848127 0.5703028 0.6852211 0.6248171 0.6540949 0.6795459
#>  [8] 0.6889107 0.6852211 0.6686914 0.6323408 0.6261037 0.6811749 0.6879248
#> [15] 0.6329467 0.6410914 0.5107526 0.5738498
#> 
#> $Joint.Entropy
#> [1] 6.818678
#> 
#> $Total.Cor
#> [1] 4.533942
#> 
#> $Normalized
#> [1] 0.4251749
#>