Weighted Topological Overlap

Computes weighted topological overlap following the Novick et al. (2009) definition

Usage

wto(network, signed = TRUE, diagonal.zero = TRUE)

Arguments

network

Symmetric matrix or data frame. A symmetric network

signed

Boolean (length = 1). Whether the signed version should be used. Defaults to TRUE. Use FALSE for absolute values

diagonal.zero

Boolean (length = 1). Whether diagonal of overlap matrix should be set to zero. Defaults to TRUE. Use FALSE to allow overlap of a node with itself

Value

A symmetric matrix of weighted topological overlap values between each pair of variables

References

Original formalization
Nowick, K., Gernat, T., Almaas, E., & Stubbs, L. (2009). Differences in human and chimpanzee gene expression patterns define an evolving network of transcription factors in brain. Proceedings of the National Academy of Sciences, 106, 22358-22363.

Examples

# Obtain network
network <- network.estimation(wmt2[,7:24], model = "glasso")

# Compute wTO
wto(network)
#>             wmt1        wmt2         wmt3       wmt4       wmt5       wmt6
#> wmt1  0.00000000 0.158996348  0.071511158 0.15119928 0.10877728 0.02766897
#> wmt2  0.15899635 0.000000000  0.249853405 0.07033602 0.15342659 0.07083681
#> wmt3  0.07151116 0.249853405  0.000000000 0.05993642 0.05691270 0.11194247
#> wmt4  0.15119928 0.070336025  0.059936423 0.00000000 0.07723982 0.07315080
#> wmt5  0.10877728 0.153426589  0.056912698 0.07723982 0.00000000 0.03680376
#> wmt6  0.02766897 0.070836811  0.111942474 0.07315080 0.03680376 0.00000000
#> wmt7  0.01977026 0.024638429  0.033841686 0.07089473 0.05097548 0.11350837
#> wmt8  0.01683998 0.029278270  0.033987630 0.05847340 0.03023955 0.09083428
#> wmt9  0.01876395 0.046732356  0.031677425 0.03431002 0.02372040 0.08016626
#> wmt10 0.06831609 0.077254636  0.070844500 0.09131439 0.04414526 0.09921954
#> wmt11 0.03303740 0.033490286  0.063056261 0.02710980 0.03430137 0.03774474
#> wmt12 0.01099310 0.028015065  0.027473828 0.02358084 0.07222493 0.03887384
#> wmt13 0.01352953 0.029422926  0.030241860 0.01878796 0.05121875 0.05386765
#> wmt14 0.01252534 0.043765516  0.017556059 0.01844815 0.02162208 0.04129679
#> wmt15 0.01340255 0.009960543  0.019912425 0.04873187 0.02154708 0.07527225
#> wmt16 0.04840764 0.024752508  0.024062316 0.02354460 0.02788301 0.04067441
#> wmt17 0.04072801 0.028333115  0.015543593 0.04025410 0.09131655 0.08774400
#> wmt18 0.01180173 0.002166914 -0.001072818 0.01095198 0.01788894 0.02610292
#>             wmt7       wmt8       wmt9      wmt10      wmt11      wmt12
#> wmt1  0.01977026 0.01683998 0.01876395 0.06831609 0.03303740 0.01099310
#> wmt2  0.02463843 0.02927827 0.04673236 0.07725464 0.03349029 0.02801506
#> wmt3  0.03384169 0.03398763 0.03167743 0.07084450 0.06305626 0.02747383
#> wmt4  0.07089473 0.05847340 0.03431002 0.09131439 0.02710980 0.02358084
#> wmt5  0.05097548 0.03023955 0.02372040 0.04414526 0.03430137 0.07222493
#> wmt6  0.11350837 0.09083428 0.08016626 0.09921954 0.03774474 0.03887384
#> wmt7  0.00000000 0.12613787 0.04386844 0.05726153 0.09426053 0.03654448
#> wmt8  0.12613787 0.00000000 0.09868954 0.03974923 0.06223273 0.08531122
#> wmt9  0.04386844 0.09868954 0.00000000 0.13353610 0.02331130 0.08921505
#> wmt10 0.05726153 0.03974923 0.13353610 0.00000000 0.06086957 0.02011850
#> wmt11 0.09426053 0.06223273 0.02331130 0.06086957 0.00000000 0.05235147
#> wmt12 0.03654448 0.08531122 0.08921505 0.02011850 0.05235147 0.00000000
#> wmt13 0.05924149 0.02348156 0.09885328 0.07613496 0.01871945 0.01985393
#> wmt14 0.08860972 0.03440631 0.13764603 0.07136262 0.02298747 0.04603490
#> wmt15 0.07868762 0.04000211 0.05982440 0.03175485 0.05190774 0.12523060
#> wmt16 0.02867420 0.04269825 0.08189031 0.07409570 0.05966700 0.02595372
#> wmt17 0.05867616 0.03366339 0.02139233 0.02149472 0.02338581 0.07734735
#> wmt18 0.03313102 0.05852820 0.02536375 0.01566573 0.05869409 0.03147499
#>            wmt13      wmt14       wmt15      wmt16      wmt17        wmt18
#> wmt1  0.01352953 0.01252534 0.013402549 0.04840764 0.04072801  0.011801733
#> wmt2  0.02942293 0.04376552 0.009960543 0.02475251 0.02833312  0.002166914
#> wmt3  0.03024186 0.01755606 0.019912425 0.02406232 0.01554359 -0.001072818
#> wmt4  0.01878796 0.01844815 0.048731872 0.02354460 0.04025410  0.010951977
#> wmt5  0.05121875 0.02162208 0.021547079 0.02788301 0.09131655  0.017888938
#> wmt6  0.05386765 0.04129679 0.075272254 0.04067441 0.08774400  0.026102922
#> wmt7  0.05924149 0.08860972 0.078687618 0.02867420 0.05867616  0.033131017
#> wmt8  0.02348156 0.03440631 0.040002112 0.04269825 0.03366339  0.058528201
#> wmt9  0.09885328 0.13764603 0.059824403 0.08189031 0.02139233  0.025363745
#> wmt10 0.07613496 0.07136262 0.031754849 0.07409570 0.02149472  0.015665728
#> wmt11 0.01871945 0.02298747 0.051907744 0.05966700 0.02338581  0.058694094
#> wmt12 0.01985393 0.04603490 0.125230603 0.02595372 0.07734735  0.031474990
#> wmt13 0.00000000 0.08723589 0.056571145 0.11499622 0.01590718  0.016898157
#> wmt14 0.08723589 0.00000000 0.100329703 0.05660288 0.04089269  0.069471875
#> wmt15 0.05657115 0.10032970 0.000000000 0.07047977 0.03731380  0.065480204
#> wmt16 0.11499622 0.05660288 0.070479766 0.00000000 0.03418579  0.086936141
#> wmt17 0.01590718 0.04089269 0.037313800 0.03418579 0.00000000  0.131731714
#> wmt18 0.01689816 0.06947188 0.065480204 0.08693614 0.13173171  0.000000000