Diagnostics Analysis for Low Stability Items

Computes item diagnostics to determine whether there exist several potential psychometric issues that affect itemStability. These issues include local dependence (using UVA), minor dimensions (using cosine on the stability patterns), multidimensional items (using net.loads), and items with low loadings (using net.loads)

Usage

itemDiagnostics(data, ...)

Arguments

data

Matrix or data frame. Should consist only of variables to be used in the analysis

...

Additional arguments to pass on to bootEGA, net.loads, and UVA

Value

Returns a list containing:

diagnostics

A data frame containing the diagnostics of low item stabilities (see itemStability)

boot

Output from bootEGA

uva

Output from UVA

minor

A list containing suggested items to keep, remove, and a matrix for probable minor dimensions (minor.matrix)

loadings

Output from net.loads

suggested

Variables that are suggested to be retained to increase item stability

Author

Alexander P. Christensen <alexpaulchristensen@gmail.com>, Hudson Golino <hfg9s at virginia.edu>, and Luis Eduardo Garrido <garrido.luiseduardo@gmail.com>

Examples

# Load data
wmt <- wmt2[,7:24]

if (FALSE) { # \dontrun{
# Obtain diagnostics
diagnostics <- itemDiagnostics(wmt, ncores = 2)} # }