Network Loadings

Computes the between- and within-community strength of each variable for each community

Usage

net.loads(
  A,
  wc,
  loading.method = c("BRM", "experimental"),
  scaling = 2,
  rotation = NULL,
  ...
)

Arguments

A

Network matrix, data frame, or EGA object

wc

Numeric or character vector (length = ncol(A)). A vector of community assignments. If input into A is an EGA object, then wc is automatically detected

loading.method

Character (length = 1). Sets network loading calculation based on implementation described in "BRM" (Christensen & Golino, 2021) or an "experimental" implementation. Defaults to "BRM"

scaling

Numeric (length = 1). Scaling factor for the magnitude of the "experimental" network loadings. Defaults to 2. 10 makes loadings roughly the size of factor loadings when correlations between factors are orthogonal

rotation

Character. A rotation to use to obtain a simpler structure. For a list of rotations, see rotations for options. Defaults to NULL or no rotation. By setting a rotation, scores estimation will be based on the rotated loadings rather than unrotated loadings

...

Additional arguments to pass on to rotations

Value

Returns a list containing:

unstd

A matrix of the unstandardized within- and between-community strength values for each node

std

A matrix of the standardized within- and between-community strength values for each node

rotated

NULL if rotation = NULL; otherwise, a list containing the rotated standardized network loadings (loadings) and correlations between dimensions (Phi) from the rotation

Details

Simulation studies have demonstrated that a node's strength centrality is roughly equivalent to factor loadings (Christensen & Golino, 2021; Hallquist, Wright, & Molenaar, 2019). Hallquist and colleagues (2019) found that node strength represented a combination of dominant and cross-factor loadings. This function computes each node's strength within each specified dimension, providing a rough equivalent to factor loadings (including cross-loadings; Christensen & Golino, 2021).

References

Original implementation and simulation
Christensen, A. P., & Golino, H. (2021). On the equivalency of factor and network loadings. Behavior Research Methods, 53, 1563-1580.

Demonstration of node strength similarity to CFA loadings
Hallquist, M., Wright, A. C. G., & Molenaar, P. C. M. (2019). Problems with centrality measures in psychopathology symptom networks: Why network psychometrics cannot escape psychometric theory. Multivariate Behavioral Research, 1-25.

Author

Alexander P. Christensen <alexpaulchristensen@gmail.com> and Hudson Golino <hfg9s at virginia.edu>

Examples

# Load data
wmt <- wmt2[,7:24]

# Estimate EGA
ega.wmt <- EGA(
  data = wmt,
  plot.EGA = FALSE # No plot for CRAN checks
)

# Network loadings
net.loads(ega.wmt)
#> Loading Method: BRM
#> 
#>       1     2    
#> wmt2  0.384      
#> wmt1  0.254      
#> wmt3  0.217 0.131
#> wmt5  0.201 0.142
#> wmt4  0.188  0.14
#> wmt9        0.293
#> wmt7        0.258
#> wmt15       0.247
#> wmt14       0.243
#> wmt6   0.14  0.24
#> wmt16        0.22
#> wmt8        0.219
#> wmt10 0.166 0.206
#> wmt12       0.187
#> wmt18       0.183
#> wmt13       0.171
#> wmt17       0.168
#> wmt11       0.155
#> Standardized loadings >= |0.10| are displayed. To change this 'minimum', use `print(net.loads_object, minimum = 0.10)`