Network Scores

This function computes network scores computed based on each node's strength within each community in the network (see net.loads). These values are used as "network loadings" for the weights of each variable.

Network scores are computed as a formative composite rather than a reflective factor. This composite representation is consistent with no latent factors that psychometric network theory proposes.

Scores can be computed as a "simple" structure, which is equivalent to a weighted sum scores or as a "full" structure, which is equivalent to an EFA approach. Conservatively, the "simple" structure approach is recommended until further validation

Usage

net.scores(
  data,
  A,
  wc,
  loading.method = c("BRM", "experimental"),
  rotation = NULL,
  scores = c("Anderson", "Bartlett", "components", "Harman", "network", "tenBerge",
    "Thurstone"),
  loading.structure = c("simple", "full"),
  impute = c("mean", "median", "none"),
  ...
)

Arguments

data

Matrix or data frame. Should consist only of variables to be used in the analysis

A

Network matrix, data frame, or EGA object

wc

Numeric or character vector (length = ncol(A)). A vector of community assignments. If input into A is an EGA object, then wc is automatically detected

loading.method

Character (length = 1). Sets network loading calculation based on implementation described in "BRM" (Christensen & Golino, 2021) or an "experimental" implementation. Defaults to "BRM"

rotation

Character. A rotation to use to obtain a simpler structure. For a list of rotations, see rotations for options. Defaults to NULL or no rotation. By setting a rotation, scores estimation will be based on the rotated loadings rather than unrotated loadings

scores

Character (length = 1). How should scores be estimated? Defaults to "network" for network scores. Set to other scoring methods which will be computed using factor.scores (see link for arguments and explanations for other methods)

loading.structure

Character (length = 1). Whether simple structure or the saturated loading matrix should be used when computing scores. Defaults to "simple"

"simple" structure more closely mirrors sum scores and CFA; "full" structure more closely mirrors EFA

Simple structure is the more "conservative" (established) approach and is therefore the default. Treat "full" as experimental as proper vetting and validation has not been established

impute

Character (length = 1). If there are any missing data, then imputation can be implemented. Available options:

"none" --- Default. No imputation is performed

"mean" --- The mean value of each variable is used to replace missing data for that variable

"median" --- The median value of each variable is used to replace missing data for that variable

...

Additional arguments to be passed on to net.loads and factor.scores

Value

Returns a list containing:

scores

A list containing the standardized (std.scores) rotated (rot.scores) scores. If rotation = NULL, then rot.scores will be NULL

loadings

Output from net.loads

References

Original implementation and simulation for loadings
Christensen, A. P., & Golino, H. (2021). On the equivalency of factor and network loadings. Behavior Research Methods, 53, 1563-1580.

Preliminary simulation for scores
Golino, H., Christensen, A. P., Moulder, R., Kim, S., & Boker, S. M. (2021). Modeling latent topics in social media using Dynamic Exploratory Graph Analysis: The case of the right-wing and left-wing trolls in the 2016 US elections. Psychometrika.

Author

Alexander P. Christensen <alexpaulchristensen@gmail.com> and Hudson F. Golino <hfg9s at virginia.edu>

Examples

# Load data
wmt <- wmt2[,7:24]

# Estimate EGA
ega.wmt <- EGA(
  data = wmt,
  plot.EGA = FALSE # No plot for CRAN checks
)

# Network scores
net.scores(data = wmt, A = ega.wmt)
#> $scores
#> $scores$std.scores
#>                   1            2
#>    [1,] -0.73602153  0.171791135
#>    [2,]  0.78840476  2.573691386
#>    [3,]  0.40975797  1.073059457
#>    [4,] -0.56696225 -1.593377368
#>    [5,]  0.78840476 -1.125767721
#>    [6,]  0.78840476 -2.015394142
#>    [7,] -1.66431630 -0.653597047
#>    [8,]  0.78840476  3.319415193
#>    [9,]  0.35701878 -0.099691629
#>   [10,]  0.40975797 -1.027281876
#>   [11,]  0.35701878  0.427376001
#>   [12,]  0.40975797  2.172096601
#>   [13,]  0.24307053 -0.587582185
#>   [14,]  0.78840476 -1.101794666
#>   [15,]  0.78840476 -1.308893959
#>   [16,] -1.54605431 -2.123545201
#>   [17,]  0.78840476 -1.182446522
#>   [18,] -0.18831545 -1.461488724
#>   [19,] -1.00072007 -1.616966601
#>   [20,]  0.78840476  2.502160525
#>   [21,]  0.78840476  3.319415193
#>   [22,] -0.13557627  1.110197508
#>   [23,]  0.78840476  1.254307561
#>   [24,] -1.06387104  0.146738511
#>   [25,]  0.78840476  0.685233754
#>   [26,] -1.11466833 -1.254677194
#>   [27,]  0.78840476 -2.117188311
#>   [28,] -2.04296310 -2.539205084
#>   [29,] -0.02162801 -0.151190606
#>   [30,]  0.78840476  2.974358129
#>   [31,]  0.78840476  2.313929505
#>   [32,] -0.18831545 -1.049530362
#>   [33,]  0.78840476  2.644294205
#>   [34,]  0.78840476 -1.616966601
#>   [35,]  0.78840476  0.848811505
#>   [36,]  0.29149597 -0.369113637
#>   [37,]  0.40975797  0.045202618
#>   [38,] -0.51853681 -1.698051688
#>   [39,]  0.40975797  1.214438597
#>   [40,] -2.04296310  0.447797990
#>   [41,]  0.78840476  1.673444184
#>   [42,] -0.51853681 -0.775422889
#>   [43,] -0.02162801 -0.200550066
#>   [44,]  0.78840476  0.304964498
#>   [45,] -0.08715083 -1.949379280
#>   [46,]  0.78840476  1.325838422
#>   [47,] -0.13557627 -1.104100505
#>   [48,]  0.78840476  0.718342265
#>   [49,]  0.40975797  2.903755310
#>   [50,] -1.00072007 -0.503329854
#>   [51,]  0.78840476  2.041205354
#>   [52,]  0.40975797 -0.941845613
#>   [53,] -1.06387104 -1.715593526
#>   [54,] -0.13989001 -0.769068755
#>   [55,]  0.40975797  0.693970926
#>   [56,] -0.02162801 -0.015439298
#>   [57,]  0.29149597 -2.015394142
#>   [58,] -0.51853681 -0.278755002
#>   [59,]  0.78840476  0.996507856
#>   [60,]  0.78840476  0.264918734
#>   [61,]  0.78840476  3.319415193
#>   [62,]  0.24307053  2.897398420
#>   [63,] -0.02162801  1.226056776
#>   [64,]  0.35701878  1.978137473
#>   [65,]  0.24307053  0.943067005
#>   [66,]  0.78840476  1.624845660
#>   [67,] -2.04296310 -2.051314529
#>   [68,]  0.40975797  2.448052960
#>   [69,]  0.40975797  0.304964498
#>   [70,] -0.13557627 -1.527503586
#>   [71,]  0.40975797  1.912334720
#>   [72,]  0.78840476  2.819193483
#>   [73,] -0.02162801 -0.840989932
#>   [74,]  0.78840476 -0.275201179
#>   [75,]  0.78840476  0.175314639
#>   [76,] -0.68759608 -2.075967797
#>   [77,] -0.56696225 -2.038983375
#>   [78,]  0.24307053  0.281285822
#>   [79,]  0.78840476  0.621378910
#>   [80,] -1.11466833  0.243321996
#>   [81,]  0.78840476  0.886356014
#>   [82,]  0.78840476 -0.485054659
#>   [83,]  0.29149597 -0.085316824
#>   [84,] -0.63248506 -2.539205084
#>   [85,]  0.24307053  0.291314986
#>   [86,]  0.40975797 -0.700199881
#>   [87,] -0.73602153  0.354906651
#>   [88,]  0.78840476  2.097743075
#>   [89,]  0.40975797  2.160313620
#>   [90,]  0.78840476  2.382226484
#>   [91,]  0.24307053 -0.117052275
#>   [92,] -0.13989001  1.208514938
#>   [93,] -0.13557627  0.825714099
#>   [94,] -0.13557627 -1.461488724
#>   [95,]  0.78840476  0.779205542
#>   [96,] -1.49762886 -1.323887099
#>   [97,]  0.78840476  1.232413666
#>   [98,]  0.35701878 -0.678916487
#>   [99,]  0.40975797  1.295481547
#>  [100,]  0.24307053  1.275059558
#>  [101,]  0.78840476  2.061451316
#>  [102,]  0.35701878  0.139704164
#>  [103,]  0.40975797 -1.141084928
#>  [104,]  0.78840476  1.054165045
#>  [105,]  0.24307053  1.170771163
#>  [106,]  0.78840476  2.573691386
#>  [107,]  0.78840476 -0.625546011
#>  [108,] -0.62207327  1.787624336
#>  [109,]  0.78840476  2.504442642
#>  [110,]  0.78840476  2.440518022
#>  [111,]  0.35701878  0.971951974
#>  [112,]  0.78840476  2.989351269
#>  [113,]  0.78840476  2.729589389
#>  [114,]  0.24307053  2.110454098
#>  [115,] -0.02162801 -1.182305442
#>  [116,]  0.78840476  3.319415193
#>  [117,]  0.40975797  1.366084366
#>  [118,]  0.29149597  1.184836262
#>  [119,] -2.04296310 -0.179242951
#>  [120,]  0.78840476  1.184836262
#>  [121,]  0.78840476  1.864757316
#>  [122,]  0.78840476  1.134002148
#>  [123,]  0.78840476  2.558698245
#>  [124,] -0.62207327  0.403697325
#>  [125,]  0.78840476  1.324387049
#>  [126,]  0.78840476  2.228634321
#>  [127,] -0.13557627  0.769176378
#>  [128,]  0.78840476  2.504442642
#>  [129,]  0.78840476  1.625545471
#>  [130,]  0.78840476  2.560980362
#>  [131,]  0.40975797 -1.185108508
#>  [132,]  0.78840476  2.906037427
#>  [133,]  0.29149597  0.557454291
#>  [134,] -0.56696225 -0.180572911
#>  [135,]  0.78840476  1.622563542
#>  [136,]  0.35701878 -0.158272789
#>  [137,]  0.78840476  0.002102540
#>  [138,] -0.08715083  2.534799517
#>  [139,]  0.78840476 -0.173265930
#>  [140,]  0.40975797  0.769176378
#>  [141,] -0.56696225 -2.123545201
#>  [142,]  0.24307053 -0.242525120
#>  [143,]  0.35701878  0.309195778
#>  [144,]  0.24307053  1.425198704
#>  [145,]  0.40975797  0.807590991
#>  [146,]  0.78840476  0.339646931
#>  [147,]  0.24307053 -0.063771242
#>  [148,] -0.02162801 -2.194148020
#>  [149,]  0.78840476  0.253961527
#>  [150,]  0.24307053  2.856177906
#>  [151,]  0.78840476  1.184836262
#>  [152,]  0.78840476  2.575973503
#>  [153,]  0.78840476  1.255835138
#>  [154,]  0.78840476  1.352019267
#>  [155,]  0.78840476  2.377450141
#>  [156,] -0.13557627  1.829321653
#>  [157,]  0.78840476  2.558698245
#>  [158,]  0.78840476 -2.015394142
#>  [159,]  0.40975797 -0.767541178
#>  [160,]  0.78840476  0.855922935
#>  [161,]  0.78840476  0.729529969
#>  [162,] -0.18831545 -0.763091735
#>  [163,]  0.78840476  1.270828279
#>  [164,]  0.78840476  2.502160525
#>  [165,] -2.04296310 -1.064266299
#>  [166,]  0.78840476  2.442800139
#>  [167,] -0.63248506 -0.108330636
#>  [168,]  0.78840476  2.160313620
#>  [169,]  0.78840476  0.766417005
#>  [170,] -0.08715083  0.611349746
#>  [171,] -0.02162801  0.264918734
#>  [172,] -0.13989001 -1.461488724
#>  [173,]  0.35701878  0.340317769
#>  [174,]  0.40975797  0.550776091
#>  [175,]  0.40975797 -0.139533589
#>  [176,] -0.02162801  1.073059457
#>  [177,]  0.78840476  0.245365436
#>  [178,]  0.78840476  0.526858949
#>  [179,]  0.78840476 -2.539205084
#>  [180,] -0.63248506 -2.038983375
#>  [181,]  0.40975797  2.002486988
#>  [182,]  0.40975797  0.940039722
#>  [183,] -0.62207327  2.027140256
#>  [184,]  0.24307053 -0.653597047
#>  [185,] -0.13557627  2.486467573
#>  [186,] -0.63248506 -1.271909536
#>  [187,]  0.78840476  1.964072375
#>  [188,]  0.78840476  2.073089807
#>  [189,]  0.78840476 -0.012945072
#>  [190,]  0.24307053 -0.092399007
#>  [191,]  0.78840476 -1.685330218
#>  [192,]  0.35701878  0.667887467
#>  [193,] -0.13557627 -1.708919451
#>  [194,]  0.78840476  3.319415193
#>  [195,]  0.78840476  2.095460958
#>  [196,]  0.40975797  1.220968757
#>  [197,]  0.78840476  2.906037427
#>  [198,]  0.24307053 -0.757881051
#>  [199,] -0.62207327  0.025781216
#>  [200,] -0.51853681 -0.841425391
#>  [201,] -0.25383827 -1.590883142
#>  [202,]  0.78840476  2.903755310
#>  [203,]  0.24307053  0.415931093
#>  [204,] -0.63248506 -0.066315548
#>  [205,]  0.78840476  2.002486988
#>  [206,] -0.56696225  1.088847478
#>  [207,]  0.40975797  1.409983008
#>  [208,] -0.13557627  0.266292682
#>  [209,] -0.02162801 -1.693926310
#>  [210,] -0.18831545 -2.015394142
#>  [211,]  0.78840476  2.575973503
#>  [212,]  0.78840476  1.476181819
#>  [213,] -0.56696225 -0.368359097
#>  [214,]  0.78840476  0.733706778
#>  [215,]  0.78840476  1.570585979
#>  [216,]  0.29149597 -1.023973552
#>  [217,]  0.78840476  2.587756484
#>  [218,] -2.04296310 -2.051314529
#>  [219,] -0.08715083 -0.498029726
#>  [220,]  0.78840476 -0.190359759
#>  [221,]  0.78840476 -0.142027815
#>  [222,] -0.68759608  0.204191450
#>  [223,]  0.78840476  2.906037427
#>  [224,] -0.18831545 -0.856249653
#>  [225,]  0.78840476  0.174144281
#>  [226,]  0.35701878 -0.165706491
#>  [227,]  0.78840476 -1.515172432
#>  [228,] -0.51853681 -0.179242951
#>  [229,]  0.78840476  0.245365436
#>  [230,]  0.78840476  2.027140256
#>  [231,]  0.78840476  2.027140256
#>  [232,]  0.78840476  1.584148930
#>  [233,]  0.78840476  0.825714099
#>  [234,] -0.08715083  1.959966853
#>  [235,]  0.40975797  2.793110025
#>  [236,] -0.02162801 -0.961267014
#>  [237,]  0.78840476  2.917820408
#>  [238,] -0.56696225 -1.090661432
#>  [239,] -1.61157712 -0.978830035
#>  [240,]  0.78840476 -0.226280146
#>  [241,] -2.04296310 -1.685330218
#>  [242,]  0.40975797  0.733706778
#>  [243,]  0.24307053  1.044182646
#>  [244,]  0.78840476  1.232413666
#>  [245,]  0.78840476 -1.245682163
#>  [246,]  0.78840476  1.645791432
#>  [247,] -0.19068729  2.549792658
#>  [248,]  0.78840476  2.418146871
#>  [249,] -0.56933409  1.600892202
#>  [250,]  0.78840476  3.319415193
#>  [251,] -0.19068729  2.020783366
#>  [252,]  0.40975797 -1.214392412
#>  [253,]  0.78840476  2.490377543
#>  [254,] -0.51853681 -1.112598243
#>  [255,] -0.25383827  0.182658898
#>  [256,]  0.78840476  0.182748341
#>  [257,]  0.78840476  2.027140256
#>  [258,]  0.40975797  0.670549453
#>  [259,] -0.56933409  0.601512818
#>  [260,] -0.63248506  1.517394412
#>  [261,]  0.40975797  2.502160525
#>  [262,] -1.61157712  2.110454098
#>  [263,] -0.08715083 -2.038983375
#>  [264,] -0.02162801 -0.823665389
#>  [265,]  0.40975797 -0.674019035
#>  [266,]  0.40975797  1.133309485
#>  [267,]  0.78840476  2.027140256
#>  [268,]  0.24307053 -1.075613821
#>  [269,] -0.08715083  0.643440566
#>  [270,]  0.78840476  1.619015310
#>  [271,]  0.78840476  1.978137473
#>  [272,]  0.78840476  2.379944367
#>  [273,] -0.02162801 -0.714170702
#>  [274,]  0.24307053 -0.186535573
#>  [275,] -0.63248506  1.532892810
#>  [276,]  0.78840476  2.174378718
#>  [277,] -0.25383827  0.661025319
#>  [278,] -1.00072007 -2.194148020
#>  [279,]  0.40975797  0.996605242
#>  [280,]  0.24307053  1.104943909
#>  [281,]  0.78840476  0.849392775
#>  [282,]  0.78840476  2.906037427
#>  [283,] -0.63248506 -1.027281876
#>  [284,]  0.78840476  2.124519197
#>  [285,] -0.19068729  3.319415193
#>  [286,]  0.78840476  2.124519197
#>  [287,]  0.24307053  0.134806712
#>  [288,]  0.78840476 -2.137610299
#>  [289,]  0.40975797  0.715280923
#>  [290,] -2.04296310 -1.379175468
#>  [291,] -0.56696225 -1.215371816
#>  [292,] -0.56696225 -0.653597047
#>  [293,] -0.56696225  1.099240302
#>  [294,] -1.16740751 -0.783513723
#>  [295,]  0.40975797  2.490377543
#>  [296,]  0.78840476 -0.052423720
#>  [297,]  0.78840476 -2.137610299
#>  [298,]  0.29149597 -0.931905892
#>  [299,] -0.02162801 -2.038983375
#>  [300,]  0.78840476  0.111128036
#>  [301,]  0.40975797  1.767679151
#>  [302,]  0.78840476  1.464066072
#>  [303,]  0.78840476 -2.015394142
#>  [304,]  0.78840476 -0.478697769
#>  [305,]  0.35701878  1.319160223
#>  [306,]  0.78840476  1.634008451
#>  [307,] -0.18831545 -1.356471363
#>  [308,]  0.29149597 -0.986009726
#>  [309,]  0.78840476  2.573691386
#>  [310,]  0.24307053 -1.243020177
#>  [311,]  0.78840476  1.109269466
#>  [312,] -0.56696225 -1.027281876
#>  [313,]  0.78840476  2.906037427
#>  [314,] -2.04296310 -1.001198418
#>  [315,]  0.40975797 -2.117188311
#>  [316,]  0.40975797 -1.027281876
#>  [317,]  0.24307053  1.352019267
#>  [318,]  0.40975797  0.337823543
#>  [319,] -0.18831545 -0.237791692
#>  [320,]  0.78840476  2.989351269
#>  [321,]  0.78840476  2.831524637
#>  [322,]  0.24307053 -0.587582185
#>  [323,]  0.78840476  0.851797557
#>  [324,] -0.13989001 -0.329510144
#>  [325,]  0.78840476  0.182748341
#>  [326,]  0.24307053  1.569155789
#>  [327,] -2.04296310 -1.007414228
#>  [328,]  0.78840476 -1.271909536
#>  [329,]  0.78840476  0.225300124
#>  [330,] -1.11898206 -1.629297755
#>  [331,]  0.29149597  2.906037427
#>  [332,]  0.24307053  0.240111836
#>  [333,]  0.24307053 -1.099512548
#>  [334,]  0.78840476  1.500835087
#>  [335,] -0.18831545 -0.752223972
#>  [336,] -0.02162801 -2.012899916
#>  [337,] -2.04296310 -1.807546375
#>  [338,] -0.18831545  0.462533926
#>  [339,] -0.02162801 -1.385529602
#>  [340,] -1.06387104 -1.551092819
#>  [341,]  0.78840476  0.397938468
#>  [342,]  0.40975797 -0.750400584
#>  [343,]  0.78840476  0.424119314
#>  [344,]  0.40975797  1.146117895
#>  [345,] -1.54605431 -2.539205084
#>  [346,] -0.56696225 -1.549662628
#>  [347,]  0.78840476 -0.165706491
#>  [348,]  0.78840476  1.425198704
#>  [349,] -0.51853681 -2.038983375
#>  [350,]  0.29149597 -0.596221192
#>  [351,] -0.18831545 -2.012899916
#>  [352,] -0.08715083 -1.787124387
#>  [353,] -0.73602153  0.001101953
#>  [354,]  0.78840476 -1.182305442
#>  [355,]  0.78840476 -1.003551564
#>  [356,]  0.78840476 -1.215371816
#>  [357,] -1.16740751 -1.337778927
#>  [358,]  0.78840476  0.714479601
#>  [359,]  0.29149597  1.270828279
#>  [360,] -1.06387104  0.253210080
#>  [361,] -0.63248506  1.319160223
#>  [362,]  0.24307053  0.199362137
#>  [363,]  0.40975797 -0.033367060
#>  [364,] -0.08715083 -1.795763394
#>  [365,] -1.61157712 -2.539205084
#>  [366,]  0.29149597  0.206473567
#>  [367,] -1.00072007  0.266097130
#>  [368,] -1.06387104 -1.109349477
#>  [369,]  0.78840476 -2.099646473
#>  [370,] -0.19068729 -0.988867263
#>  [371,] -0.56696225 -1.305590721
#>  [372,]  0.78840476 -1.269360839
#>  [373,] -1.66431630 -2.117188311
#>  [374,]  0.78840476  1.101034717
#>  [375,] -1.11466833 -2.539205084
#>  [376,] -1.00072007 -2.539205084
#>  [377,]  0.40975797 -1.693926310
#>  [378,] -0.13557627 -0.922852401
#>  [379,]  0.78840476  1.454326531
#>  [380,]  0.24307053 -0.662236054
#>  [381,] -1.61157712 -1.105486813
#>  [382,] -0.13557627 -0.625687091
#>  [383,] -2.04296310 -2.539205084
#>  [384,] -2.04296310 -0.922852401
#>  [385,]  0.78840476 -0.063771242
#>  [386,]  0.35701878 -1.170115367
#>  [387,] -0.08715083  2.027140256
#>  [388,] -1.54605431 -1.623323491
#>  [389,] -1.06387104 -1.604322215
#>  [390,] -1.54605431 -2.015394142
#>  [391,] -1.06624288 -2.038983375
#>  [392,]  0.24307053 -1.629297755
#>  [393,]  0.35701878 -1.090661432
#>  [394,]  0.40975797  1.012944531
#>  [395,] -0.68522425 -2.099646473
#>  [396,] -0.02162801  0.738294735
#>  [397,] -2.04296310 -2.209141161
#>  [398,] -2.04296310 -1.198139473
#>  [399,]  0.78840476  0.968741814
#>  [400,] -1.23293032 -1.308572649
#>  [401,]  0.35701878 -0.151190606
#>  [402,]  0.29149597 -0.765046952
#>  [403,] -1.66431630 -0.937677781
#>  [404,]  0.35701878 -1.512678206
#>  [405,] -1.49762886 -2.123545201
#>  [406,] -1.23293032 -1.623323491
#>  [407,] -0.02162801 -1.599424763
#>  [408,]  0.29149597 -1.286902677
#>  [409,]  0.78840476 -1.215371816
#>  [410,] -1.00072007 -1.693926310
#>  [411,] -2.04296310 -0.672144011
#>  [412,] -1.06387104 -2.539205084
#>  [413,]  0.78840476  1.500835087
#>  [414,]  0.78840476 -0.563992952
#>  [415,]  0.40975797 -1.949379280
#>  [416,] -0.68522425 -1.075613821
#>  [417,]  0.78840476 -0.925346627
#>  [418,] -1.54605431 -2.539205084
#>  [419,] -2.04296310 -1.111843703
#>  [420,] -2.04296310 -1.745903873
#>  [421,] -0.08715083 -1.949379280
#>  [422,] -1.16740751 -1.619315356
#>  [423,] -2.04296310 -1.575835531
#>  [424,] -2.04296310 -0.549308652
#>  [425,] -1.54605431 -2.539205084
#>  [426,]  0.78840476 -0.289127954
#>  [427,] -1.54605431 -2.123545201
#>  [428,] -1.11898206 -1.186356492
#>  [429,] -1.66431630 -1.164763604
#>  [430,] -1.06387104 -1.613799357
#>  [431,] -2.04296310 -1.527503586
#>  [432,] -1.06387104 -2.117188311
#>  [433,]  0.78840476  1.142907736
#>  [434,] -1.16740751 -1.164763604
#>  [435,] -1.61157712 -2.539205084
#>  [436,] -1.66431630 -1.772131246
#>  [437,] -0.63248506 -1.604322215
#>  [438,] -1.11466833 -1.637936762
#>  [439,] -1.49762886 -2.038983375
#>  [440,] -0.51853681 -1.125767721
#>  [441,] -0.25383827 -2.539205084
#>  [442,] -1.54605431 -2.539205084
#>  [443,] -0.68522425 -1.202727430
#>  [444,] -1.06387104 -2.123545201
#>  [445,]  0.40975797 -1.201306717
#>  [446,] -2.04296310 -2.194148020
#>  [447,] -0.68759608 -1.110104017
#>  [448,] -2.04296310 -1.685330218
#>  [449,] -1.49762886 -0.267319468
#>  [450,]  0.40975797  0.229672484
#>  [451,] -2.04296310 -1.616966601
#>  [452,] -1.49762886 -1.108225883
#>  [453,] -0.13989001 -1.104100505
#>  [454,] -1.61157712 -2.099646473
#>  [455,]  0.35701878 -0.763091735
#>  [456,] -0.56696225 -1.649719743
#>  [457,]  0.40975797  1.219605256
#>  [458,]  0.40975797  2.793110025
#>  [459,] -0.02162801 -0.233839585
#>  [460,] -1.66431630 -2.209141161
#>  [461,]  0.78840476  2.034675194
#>  [462,] -1.54605431 -0.829094237
#>  [463,] -0.56696225 -2.038983375
#>  [464,] -0.56696225 -2.075967797
#>  [465,] -0.02162801 -0.750447349
#>  [466,]  0.78840476 -0.299854447
#>  [467,]  0.78840476  1.452283091
#>  [468,]  0.78840476  0.009192787
#>  [469,] -1.61157712 -1.685330218
#>  [470,]  0.35701878 -1.527362506
#>  [471,]  0.40975797 -0.629918371
#>  [472,] -2.04296310 -1.693926310
#>  [473,] -1.06387104 -1.442067322
#>  [474,] -1.49762886 -0.893234075
#>  [475,] -0.63248506 -2.194148020
#>  [476,] -0.56933409 -2.539205084
#>  [477,] -0.08715083 -1.686268707
#>  [478,] -1.61157712 -2.194148020
#>  [479,]  0.78840476  1.634008451
#>  [480,]  0.40975797 -1.708919451
#>  [481,] -1.11466833 -1.619315356
#>  [482,] -2.04296310 -1.160175647
#>  [483,]  0.78840476 -0.656000273
#>  [484,]  0.35701878 -0.873524911
#>  [485,]  0.35701878  0.577035388
#>  [486,]  0.35701878 -0.269849415
#>  [487,]  0.78840476 -0.745549897
#>  [488,] -0.13557627  1.100630459
#>  [489,] -2.04296310 -2.123545201
#>  [490,] -0.68522425 -1.637388589
#>  [491,]  0.40975797 -1.551092819
#>  [492,]  0.29149597 -1.075613821
#>  [493,]  0.78840476  0.289787409
#>  [494,]  0.78840476 -0.844231293
#>  [495,] -1.61157712 -1.198139473
#>  [496,] -0.62207327 -1.144031909
#>  [497,]  0.29149597 -0.780710656
#>  [498,]  0.29149597 -0.601956778
#>  [499,] -2.04296310 -0.985920282
#>  [500,] -2.04296310 -0.221793425
#>  [501,]  0.40975797 -2.209141161
#>  [502,]  0.35701878 -0.242666200
#>  [503,]  0.35701878 -0.310888737
#>  [504,] -0.19068729  1.352019267
#>  [505,]  0.40975797  0.056824921
#>  [506,]  0.78840476 -0.827360293
#>  [507,] -0.13557627  2.136414892
#>  [508,]  0.78840476  2.490377543
#>  [509,]  0.78840476  1.658184463
#>  [510,] -0.08715083 -1.677629699
#>  [511,]  0.35701878  0.634844474
#>  [512,]  0.78840476  2.119139634
#>  [513,] -0.02162801  1.743725694
#>  [514,]  0.78840476 -1.093155658
#>  [515,]  0.40975797 -1.153818757
#>  [516,] -0.08715083 -0.769448625
#>  [517,]  0.78840476  0.047696844
#>  [518,] -2.04296310 -1.125908801
#>  [519,] -2.04296310 -2.539205084
#>  [520,] -1.23293032 -1.593377368
#>  [521,]  0.78840476 -0.659953937
#>  [522,] -0.56933409 -1.174240745
#>  [523,] -1.54605431 -2.051314529
#>  [524,] -1.61157712 -1.793481277
#>  [525,]  0.78840476 -0.329890013
#>  [526,] -1.06387104 -1.095504413
#>  [527,] -1.61157712 -0.523751842
#>  [528,]  0.78840476 -0.323533123
#>  [529,] -2.04296310  0.171791135
#>  [530,] -0.13557627  1.954909584
#>  [531,]  0.78840476  0.155018614
#>  [532,] -0.63248506 -0.763168819
#>  [533,] -1.66431630 -0.303717111
#>  [534,] -1.06624288 -1.153818757
#>  [535,]  0.78840476  1.999808813
#>  [536,] -0.13557627  0.703349124
#>  [537,]  0.78840476  2.906037427
#>  [538,] -0.63248506 -1.330243989
#>  [539,]  0.78840476  1.027009630
#>  [540,]  0.78840476  2.379732258
#>  [541,]  0.78840476  0.084415795
#>  [542,]  0.78840476  1.928277952
#>  [543,] -1.06624288 -0.738158873
#>  [544,]  0.40975797  1.144994302
#>  [545,] -1.61157712  0.835500947
#>  [546,]  0.78840476 -1.599424763
#>  [547,] -0.08715083 -2.539205084
#>  [548,]  0.78840476 -0.592620717
#>  [549,]  0.24307053  0.537465533
#>  [550,]  0.78840476 -2.539205084
#>  [551,] -0.19068729  0.905349226
#>  [552,] -0.73602153 -1.590883142
#>  [553,] -0.56696225 -1.090661432
#>  [554,]  0.78840476 -1.286902677
#>  [555,] -0.13557627 -0.629772900
#>  [556,] -0.02162801 -1.515172432
#>  [557,] -0.13557627 -0.612473919
#>  [558,] -0.13557627 -2.194148020
#>  [559,] -0.13557627 -2.539205084
#>  [560,] -0.56696225 -0.158272789
#>  [561,]  0.35701878 -1.662590031
#>  [562,] -0.19068729  0.314136924
#>  [563,]  0.78840476 -2.012899916
#>  [564,] -0.19068729 -0.697705654
#>  [565,] -2.04296310 -1.653951023
#>  [566,]  0.78840476  0.254050971
#>  [567,]  0.78840476 -0.674116422
#>  [568,] -0.19068729  0.485515769
#>  [569,] -2.04296310 -1.525009360
#>  [570,]  0.78840476 -0.634185018
#>  [571,] -0.19068729  0.314136924
#>  [572,]  0.78840476  1.705761867
#>  [573,] -0.13557627 -0.271892855
#>  [574,]  0.78840476  0.955431256
#>  [575,]  0.78840476  0.535497956
#>  [576,]  0.24307053  0.314136924
#>  [577,]  0.78840476  0.314136924
#>  [578,] -1.00072007 -1.949379280
#>  [579,] -0.08715083 -2.051314529
#>  [580,]  0.40975797  0.661530577
#>  [581,]  0.78840476  0.754183237
#>  [582,]  0.78840476 -1.611755917
#>  [583,]  0.78840476 -0.389004230
#>  [584,]  0.78840476  2.153956730
#>  [585,]  0.78840476  2.070807689
#>  [586,] -1.16740751 -0.694414886
#>  [587,]  0.78840476 -2.117188311
#>  [588,]  0.40975797 -0.462109340
#>  [589,] -0.02162801 -2.038983375
#>  [590,]  0.40975797 -1.515172432
#>  [591,] -0.02162801 -1.099512548
#>  [592,]  0.78840476  0.510865871
#>  [593,] -0.56933409 -1.575835531
#>  [594,]  0.35701878  0.750184580
#>  [595,]  0.78840476  2.573691386
#>  [596,] -2.04296310 -2.539205084
#>  [597,]  0.78840476  0.550635011
#>  [598,] -0.68522425 -1.527362506
#>  [599,] -0.13557627  0.220712167
#>  [600,]  0.78840476 -0.257518261
#>  [601,] -0.19068729  0.281285822
#>  [602,] -0.56696225 -1.949379280
#>  [603,] -0.13557627 -0.089904781
#>  [604,]  0.78840476  0.378061544
#>  [605,] -1.06387104 -2.117188311
#>  [606,]  0.78840476  2.552341355
#>  [607,] -0.63248506 -1.527362506
#>  [608,] -0.13557627  0.928382705
#>  [609,]  0.78840476 -0.215679200
#>  [610,] -1.06624288 -1.674373012
#>  [611,] -0.56696225 -0.757791607
#>  [612,]  0.78840476  1.302919098
#>  [613,]  0.24307053  1.206883786
#>  [614,]  0.78840476 -0.961267014
#>  [615,] -1.54605431 -1.787124387
#>  [616,]  0.40975797 -2.051314529
#>  [617,] -2.04296310 -2.075967797
#>  [618,] -1.06387104  0.004596766
#>  [619,] -1.06387104  0.004596766
#>  [620,]  0.40975797 -1.170115367
#>  [621,]  0.78840476  1.975855356
#>  [622,] -0.13557627 -2.539205084
#>  [623,]  0.40975797 -1.619315356
#>  [624,]  0.29149597  2.016552086
#>  [625,]  0.78840476  0.769176378
#>  [626,]  0.40975797  1.107250527
#>  [627,]  0.40975797 -0.201176091
#>  [628,] -0.25383827 -0.114066223
#>  [629,]  0.78840476  0.950424482
#>  [630,]  0.78840476  0.329617767
#>  [631,]  0.78840476  0.933439975
#>  [632,] -0.08715083  0.012884934
#>  [633,] -2.04296310 -2.075967797
#>  [634,] -1.49762886 -2.099646473
#>  [635,]  0.29149597  2.484020653
#>  [636,]  0.35701878 -2.099646473
#>  [637,] -0.13557627  0.107119901
#>  [638,] -0.63248506 -1.683986589
#>  [639,]  0.78840476  2.906037427
#>  [640,]  0.78840476  1.607827606
#>  [641,]  0.78840476 -0.823665389
#>  [642,] -2.04296310 -1.003551564
#>  [643,]  0.78840476  1.630098480
#>  [644,] -0.56696225  2.174378718
#>  [645,] -1.49762886 -1.730910732
#>  [646,]  0.40975797 -1.059893939
#>  [647,] -0.19068729  1.098540491
#>  [648,]  0.78840476  0.115805437
#>  [649,]  0.78840476  2.368287350
#>  [650,] -1.00072007 -2.539205084
#>  [651,]  0.78840476  1.975855356
#>  [652,] -1.23293032 -1.551092819
#>  [653,] -0.13989001 -2.194148020
#>  [654,] -1.06387104 -0.682083732
#>  [655,]  0.78840476 -0.361496949
#>  [656,]  0.78840476  0.170417187
#>  [657,] -0.56696225 -1.461488724
#>  [658,] -2.04296310 -0.483293790
#>  [659,] -0.08715083 -2.539205084
#>  [660,] -0.02162801 -1.254367699
#>  [661,]  0.29149597 -1.515172432
#>  [662,]  0.40975797  2.490377543
#>  [663,] -0.13557627  1.145569722
#>  [664,]  0.78840476  2.050818932
#>  [665,]  0.78840476  2.145320479
#>  [666,] -1.54605431 -1.667842851
#>  [667,] -2.04296310 -2.539205084
#>  [668,] -2.04296310 -2.012899916
#>  [669,]  0.78840476  1.194192635
#>  [670,]  0.40975797  0.424119314
#>  [671,] -0.13557627  2.504442642
#>  [672,] -0.56696225 -0.534824736
#>  [673,] -0.08715083  0.592661702
#>  [674,] -0.18831545 -0.314896872
#>  [675,]  0.35701878 -1.616966601
#>  [676,] -1.61157712 -2.137610299
#>  [677,] -2.04296310 -0.616209949
#>  [678,] -1.49762886 -1.449157570
#>  [679,]  0.40975797 -1.593377368
#>  [680,] -1.06624288 -1.635654645
#>  [681,]  0.78840476  2.906037427
#>  [682,] -1.54605431 -0.503329854
#>  [683,] -1.11466833 -1.177407989
#>  [684,]  0.40975797 -1.104100505
#>  [685,]  0.78840476  0.509414497
#>  [686,]  0.29149597 -0.694414886
#>  [687,]  0.35701878  1.743725694
#>  [688,]  0.24307053  0.646960221
#>  [689,] -0.08715083 -1.662590031
#>  [690,] -0.56696225 -1.949379280
#>  [691,] -1.66431630 -1.599424763
#>  [692,]  0.40975797  2.379732258
#>  [693,] -0.63248506  0.250404179
#>  [694,] -0.13557627  0.281285822
#>  [695,] -2.04296310 -2.038983375
#>  [696,]  0.78840476  0.913440060
#>  [697,] -1.00072007 -0.439395147
#>  [698,] -0.02162801 -0.602770877
#>  [699,]  0.24307053  0.368032379
#>  [700,] -0.13557627  2.034675194
#>  [701,]  0.40975797  0.087538804
#>  [702,]  0.24307053  3.319415193
#>  [703,]  0.78840476  0.855922935
#>  [704,]  0.29149597  0.355701225
#>  [705,]  0.40975797 -2.209141161
#>  [706,] -0.02162801 -2.012899916
#>  [707,] -0.68759608 -1.682835992
#>  [708,]  0.35701878  1.193521797
#>  [709,] -0.62207327  0.863457873
#>  [710,] -0.62207327 -0.978830035
#>  [711,]  0.78840476  0.488009995
#>  [712,]  0.78840476  0.573385920
#>  [713,] -1.06624288 -2.123545201
#>  [714,] -1.11466833 -2.539205084
#>  [715,] -1.49762886 -1.164763604
#>  [716,] -0.02162801 -1.949379280
#>  [717,] -0.13557627 -0.063771242
#>  [718,]  0.78840476  1.546784638
#>  [719,]  0.78840476 -1.112598243
#>  [720,]  0.78840476  0.596161054
#>  [721,] -0.18831545 -0.194193176
#>  [722,]  0.24307053 -2.209141161
#>  [723,]  0.35701878 -1.263313445
#>  [724,]  0.78840476  2.088782758
#>  [725,] -0.18831545  1.319160223
#>  [726,]  0.24307053  1.257517720
#>  [727,]  0.24307053 -1.706257464
#>  [728,] -0.08715083 -1.141084928
#>  [729,] -0.56696225 -1.182305442
#>  [730,]  0.40975797 -2.099646473
#>  [731,] -0.13557627 -0.651412316
#>  [732,] -0.13557627  0.266292682
#>  [733,] -0.56696225 -1.575835531
#>  [734,]  0.78840476  1.269300702
#>  [735,]  0.78840476  3.319415193
#>  [736,]  0.24307053  1.232413666
#>  [737,]  0.78840476 -2.099646473
#>  [738,]  0.78840476 -0.690722739
#>  [739,] -0.13557627 -1.629297755
#>  [740,]  0.78840476 -0.012945072
#>  [741,]  0.78840476  1.705761867
#>  [742,]  0.40975797 -1.864084096
#>  [743,] -0.02162801 -2.539205084
#>  [744,]  0.78840476  0.387134891
#>  [745,]  0.29149597 -1.276034914
#>  [746,] -1.49762886 -2.539205084
#>  [747,]  0.40975797  0.754183237
#>  [748,]  0.40975797 -2.194148020
#>  [749,] -0.56696225 -2.539205084
#>  [750,] -1.06624288 -2.539205084
#>  [751,] -1.49762886 -2.539205084
#>  [752,] -0.25383827 -1.629297755
#>  [753,]  0.40975797  0.208381013
#>  [754,] -1.54605431 -2.051314529
#>  [755,] -1.23293032 -2.539205084
#>  [756,] -0.02162801  0.803555057
#>  [757,]  0.78840476  0.747017336
#>  [758,] -1.06624288 -0.559672229
#>  [759,]  0.29149597 -0.148243625
#>  [760,] -0.18831545 -1.129076045
#>  [761,]  0.78840476  2.050818932
#>  [762,]  0.78840476  2.526113982
#>  [763,]  0.24307053 -0.586293074
#>  [764,] -1.00072007 -0.399041459
#>  [765,] -0.51853681 -0.605265103
#>  [766,] -1.16740751 -1.425568337
#>  [767,]  0.78840476  1.544306970
#>  [768,]  0.24307053 -0.869335347
#>  [769,]  0.35701878  3.319415193
#>  [770,]  0.78840476  0.547829110
#>  [771,]  0.78840476  1.978137473
#>  [772,] -0.08715083 -0.170603943
#>  [773,] -0.63248506 -0.125872474
#>  [774,]  0.78840476  1.074587034
#>  [775,]  0.35701878 -1.203588835
#>  [776,] -0.56696225 -1.177717485
#>  [777,] -1.00072007 -1.682835992
#>  [778,]  0.78840476  1.073059457
#>  [779,] -1.11466833 -2.015394142
#>  [780,] -0.63248506  0.792855054
#>  [781,]  0.40975797 -1.093155658
#>  [782,] -0.02162801  0.777861914
#>  [783,] -0.02162801 -0.181904937
#>  [784,]  0.40975797 -0.679777892
#>  [785,]  0.35701878 -0.620335327
#>  [786,] -1.61157712 -2.539205084
#>  [787,] -1.49762886 -0.572542516
#>  [788,]  0.40975797 -0.559672229
#>  [789,]  0.35701878 -0.849935678
#>  [790,]  0.78840476  2.047386217
#>  [791,] -0.56933409 -0.112807963
#>  [792,]  0.78840476  2.097743075
#>  [793,]  0.78840476  2.145320479
#>  [794,]  0.78840476  2.558698245
#>  [795,]  0.29149597  1.235413150
#>  [796,] -0.56933409 -1.283735433
#>  [797,]  0.35701878  0.667887467
#>  [798,]  0.78840476 -0.295623167
#>  [799,]  0.78840476  2.454583120
#>  [800,] -1.61157712 -0.874541640
#>  [801,]  0.40975797 -1.593377368
#>  [802,]  0.78840476  1.966692565
#>  [803,]  0.78840476 -0.137655455
#>  [804,]  0.29149597  2.027140256
#>  [805,] -0.02162801 -1.177717485
#>  [806,]  0.40975797  1.952627466
#>  [807,] -0.08715083 -1.009598958
#>  [808,]  0.35701878 -1.144031909
#>  [809,] -0.56696225 -2.117188311
#>  [810,]  0.40975797 -0.081699004
#>  [811,]  0.29149597 -1.116431659
#>  [812,]  0.78840476  0.310175182
#>  [813,] -0.02162801 -2.539205084
#>  [814,] -0.56696225 -2.015394142
#>  [815,]  0.78840476  1.978137473
#>  [816,]  0.78840476  3.319415193
#>  [817,]  0.40975797  0.645934792
#>  [818,] -0.02162801  0.155718425
#>  [819,]  0.78840476  3.319415193
#>  [820,]  0.78840476  1.500835087
#>  [821,]  0.40975797  2.160313620
#>  [822,]  0.78840476  2.974358129
#>  [823,] -0.13557627  2.043523553
#>  [824,] -0.13557627  1.520628293
#>  [825,]  0.40975797  0.229672484
#>  [826,]  0.40975797 -0.873524911
#>  [827,]  0.35701878  2.572763344
#>  [828,]  0.78840476  2.903755310
#>  [829,]  0.78840476  0.208381013
#>  [830,]  0.78840476  2.793110025
#>  [831,]  0.78840476  1.467542812
#>  [832,]  0.78840476  1.208514938
#>  [833,]  0.78840476  2.088782758
#>  [834,]  0.78840476  1.336702061
#>  [835,]  0.78840476  1.975155545
#>  [836,]  0.78840476  1.900551739
#>  [837,]  0.78840476  0.164357433
#>  [838,]  0.29149597  1.295481547
#>  [839,]  0.29149597  1.199249905
#>  [840,]  0.78840476  3.319415193
#>  [841,]  0.78840476  2.504442642
#>  [842,]  0.40975797 -2.209141161
#>  [843,]  0.78840476 -1.271909536
#>  [844,]  0.35701878  2.454583120
#>  [845,] -0.13557627 -2.038983375
#>  [846,]  0.78840476  1.682083191
#>  [847,] -1.06387104 -0.459615114
#>  [848,] -0.68522425 -0.748098594
#>  [849,] -0.13557627  1.538578862
#>  [850,] -0.25383827 -1.203655472
#>  [851,]  0.24307053  1.193521797
#>  [852,]  0.78840476 -0.679777892
#>  [853,]  0.78840476  1.440261432
#>  [854,]  0.40975797  0.378061544
#>  [855,] -0.68759608  0.289787409
#>  [856,]  0.78840476 -1.009598958
#>  [857,] -0.51853681 -0.165706491
#>  [858,]  0.29149597 -0.605265103
#>  [859,]  0.78840476  2.014269969
#>  [860,] -2.04296310 -2.075967797
#>  [861,]  0.78840476  1.194449839
#>  [862,]  0.78840476  0.604624035
#>  [863,]  0.78840476  1.436351462
#>  [864,]  0.78840476  0.731212552
#>  [865,]  0.35701878 -0.218846444
#>  [866,] -0.13557627 -0.063771242
#>  [867,] -0.08715083  0.875476234
#>  [868,]  0.40975797 -1.721950416
#>  [869,]  0.29149597 -0.063868629
#>  [870,]  0.29149597 -0.605265103
#>  [871,] -1.66431630 -0.620119920
#>  [872,]  0.35701878  1.413661770
#>  [873,] -0.68522425 -0.498119170
#>  [874,]  0.35701878  1.654751749
#>  [875,]  0.29149597  0.223772547
#>  [876,]  0.29149597 -0.198244226
#>  [877,]  0.24307053  0.343249634
#>  [878,] -0.13989001  0.339557487
#>  [879,]  0.29149597  0.229124311
#>  [880,]  0.29149597 -0.108640132
#>  [881,] -1.11898206  1.253156963
#>  [882,] -2.04296310 -1.547784495
#>  [883,] -1.61157712 -1.114125820
#>  [884,]  0.78840476  2.903755310
#>  [885,]  0.78840476  2.088782758
#>  [886,] -1.16740751 -0.780710656
#>  [887,] -0.13557627 -1.283735433
#>  [888,]  0.78840476 -0.295482087
#>  [889,]  0.78840476  1.675726301
#>  [890,]  0.40975797  1.166861193
#>  [891,]  0.78840476  1.964072375
#>  [892,] -0.02162801 -2.539205084
#>  [893,]  0.40975797 -1.003551564
#>  [894,] -2.04296310 -2.539205084
#>  [895,] -2.04296310 -1.260819218
#>  [896,] -2.04296310 -2.539205084
#>  [897,]  0.40975797 -0.454388288
#>  [898,]  0.40975797 -0.176651338
#>  [899,]  0.78840476  2.046458175
#>  [900,] -0.02162801 -1.730910732
#>  [901,] -0.13557627 -0.768520582
#>  [902,] -0.19068729 -0.060514554
#>  [903,]  0.78840476  1.271582819
#>  [904,] -0.02162801  1.075341574
#>  [905,] -0.56696225 -1.701528427
#>  [906,]  0.78840476  0.266292682
#>  [907,] -1.11466833  1.526918546
#>  [908,]  0.40975797 -0.437456071
#>  [909,]  0.40975797 -1.213637872
#>  [910,]  0.78840476  1.574711357
#>  [911,]  0.78840476  1.319160223
#>  [912,] -0.02162801 -0.823754833
#>  [913,]  0.78840476  1.758718834
#>  [914,]  0.29149597 -1.575746087
#>  [915,] -0.73602153 -1.599424763
#>  [916,]  0.78840476  1.789906454
#>  [917,]  0.40975797  0.667887467
#>  [918,]  0.78840476 -0.124035191
#>  [919,] -0.56696225 -1.754589408
#>  [920,] -0.56696225 -1.108225883
#>  [921,] -1.06387104 -2.539205084
#>  [922,]  0.35701878  1.145189853
#>  [923,] -0.56696225  0.792855054
#>  [924,] -1.23293032  0.304964498
#>  [925,] -0.62207327  1.145189853
#>  [926,] -2.04296310 -1.533719396
#>  [927,] -1.06387104 -0.811021004
#>  [928,] -0.19068729 -0.925346627
#>  [929,] -0.02162801 -0.170912784
#>  [930,]  0.78840476  2.050818932
#>  [931,]  0.78840476 -0.063771242
#>  [932,]  0.78840476  1.460993127
#>  [933,] -0.18831545  0.729529969
#>  [934,]  0.78840476 -0.006030851
#>  [935,]  0.24307053 -1.162457764
#>  [936,] -0.56696225  0.243321996
#>  [937,]  0.40975797  0.949423895
#>  [938,]  0.78840476  2.879856582
#>  [939,]  0.78840476 -0.799012121
#>  [940,]  0.40975797 -0.800076157
#>  [941,]  0.35701878 -0.459615114
#>  [942,]  0.24307053  0.331047957
#>  [943,]  0.78840476  0.807590991
#>  [944,] -0.56696225 -0.925346627
#>  [945,] -0.63248506 -1.182305442
#>  [946,]  0.29149597 -0.682083732
#>  [947,]  0.78840476  1.266095511
#>  [948,]  0.35701878  1.730694729
#>  [949,]  0.40975797 -0.748098594
#>  [950,]  0.78840476  2.989351269
#>  [951,]  0.78840476  0.179801360
#>  [952,]  0.78840476  0.855922935
#>  [953,]  0.78840476  2.490377543
#>  [954,] -0.08715083  0.153103091
#>  [955,] -1.66431630 -1.599424763
#>  [956,] -1.49762886 -1.527362506
#>  [957,] -2.04296310 -1.772131246
#>  [958,]  0.78840476  1.500835087
#>  [959,]  0.78840476 -0.474440494
#>  [960,] -1.06387104 -1.039471950
#>  [961,]  0.78840476  2.037169420
#>  [962,]  0.78840476  2.587756484
#>  [963,]  0.78840476  2.160313620
#>  [964,] -0.08715083 -0.334720828
#>  [965,]  0.35701878  1.194192635
#>  [966,]  0.78840476  2.903755310
#>  [967,] -1.49762886 -2.539205084
#>  [968,] -1.54605431 -1.304662679
#>  [969,]  0.35701878 -0.738158873
#>  [970,]  0.78840476  0.220571087
#>  [971,] -2.04296310 -2.539205084
#>  [972,]  0.24307053 -0.156972076
#>  [973,] -0.02162801 -1.686268707
#>  [974,]  0.78840476 -0.751961258
#>  [975,]  0.40975797 -0.526185729
#>  [976,]  0.29149597 -0.477246395
#>  [977,]  0.78840476  2.974358129
#>  [978,] -0.13989001 -1.263313445
#>  [979,] -2.04296310 -2.539205084
#>  [980,]  0.78840476 -0.308539982
#>  [981,]  0.78840476  1.185986859
#>  [982,]  0.78840476  1.609852519
#>  [983,] -0.02162801 -0.599913339
#>  [984,]  0.29149597 -0.697076872
#>  [985,]  0.40975797 -0.163212265
#>  [986,]  0.40975797 -0.326077429
#>  [987,]  0.78840476  1.185986859
#>  [988,]  0.40975797  0.275722311
#>  [989,]  0.78840476 -0.653597047
#>  [990,]  0.78840476  1.716845038
#>  [991,] -2.04296310 -2.539205084
#>  [992,]  0.35701878 -1.271909536
#>  [993,]  0.78840476 -0.539250240
#>  [994,]  0.29149597 -1.593377368
#>  [995,] -0.13557627  0.792855054
#>  [996,] -1.06624288 -2.117188311
#>  [997,] -0.56696225 -2.015394142
#>  [998,] -0.73602153  0.220712167
#>  [999,]  0.78840476  2.575973503
#> [1000,] -0.51853681 -2.038983375
#> [1001,] -0.68522425 -2.038983375
#> [1002,]  0.78840476  0.792855054
#> [1003,]  0.78840476  0.769176378
#> [1004,]  0.78840476  2.061451316
#> [1005,] -1.06387104 -0.749103720
#> [1006,] -2.04296310 -2.137610299
#> [1007,] -0.08715083  1.319160223
#> [1008,]  0.35701878  1.697076332
#> [1009,]  0.24307053 -0.625546011
#> [1010,] -0.13557627 -2.117188311
#> [1011,]  0.35701878 -0.123590357
#> [1012,]  0.29149597 -0.662236054
#> [1013,] -0.13557627  1.052637468
#> [1014,]  0.78840476  2.003241528
#> [1015,]  0.78840476  2.003241528
#> [1016,] -0.18831545 -0.689066647
#> [1017,]  0.78840476  2.989351269
#> [1018,]  0.78840476 -2.051314529
#> [1019,]  0.78840476  0.991540030
#> [1020,]  0.78840476  1.649224147
#> [1021,] -1.61157712 -1.599424763
#> [1022,]  0.29149597 -2.539205084
#> [1023,]  0.78840476  2.903755310
#> [1024,] -0.13557627 -0.233929028
#> [1025,] -0.13557627  1.413661770
#> [1026,]  0.29149597  0.503754440
#> [1027,] -2.04296310 -1.730910732
#> [1028,] -0.13557627  1.015653046
#> [1029,]  0.35701878  0.068224734
#> [1030,]  0.78840476  2.047386217
#> [1031,] -1.49762886 -1.134406728
#> [1032,]  0.78840476 -0.063771242
#> [1033,]  0.78840476  2.644294205
#> [1034,] -1.00072007 -1.670337077
#> [1035,]  0.78840476  2.088782758
#> [1036,]  0.78840476 -0.015439298
#> [1037,]  0.29149597 -1.075613821
#> [1038,]  0.40975797 -2.539205084
#> [1039,]  0.40975797  1.303016485
#> [1040,]  0.35701878  0.230823082
#> [1041,]  0.35701878  0.661111881
#> [1042,] -0.02162801 -0.688087243
#> [1043,]  0.78840476  0.716219411
#> [1044,]  0.78840476  1.806617548
#> [1045,] -0.68759608  0.377920464
#> [1046,]  0.40975797  0.684703995
#> [1047,]  0.40975797  1.666765985
#> [1048,]  0.78840476  2.587756484
#> [1049,] -0.13557627 -1.613799357
#> [1050,]  0.78840476  1.684206045
#> [1051,] -2.04296310 -1.153818757
#> [1052,] -1.49762886 -1.179756744
#> [1053,] -2.04296310 -1.210161132
#> [1054,]  0.29149597  2.003241528
#> [1055,] -0.18831545 -0.515661008
#> [1056,]  0.78840476  1.564272005
#> [1057,]  0.78840476  2.572763344
#> [1058,]  0.35701878  1.216720714
#> [1059,] -0.51853681 -2.038983375
#> [1060,]  0.40975797 -0.142027815
#> [1061,] -0.19068729  1.241586091
#> [1062,]  0.24307053  1.099240302
#> [1063,]  0.40975797 -0.292820101
#> [1064,]  0.24307053  2.064450799
#> [1065,]  0.35701878 -1.792553235
#> [1066,] -0.13557627 -1.619315356
#> [1067,]  0.78840476 -0.748098594
#> [1068,]  0.78840476 -0.242525120
#> [1069,]  0.78840476  2.587756484
#> [1070,] -2.04296310 -1.210161132
#> [1071,]  0.35701878  0.230823082
#> [1072,] -1.66431630 -1.075613821
#> [1073,]  0.78840476  2.903755310
#> [1074,]  0.35701878  0.281285822
#> [1075,]  0.78840476 -0.483603285
#> [1076,]  0.78840476  2.502160525
#> [1077,] -0.02162801 -2.117188311
#> [1078,]  0.40975797 -1.683986589
#> [1079,]  0.40975797  1.660307317
#> [1080,]  0.78840476  0.950424482
#> [1081,] -0.02162801 -0.136676051
#> [1082,]  0.78840476  1.243868208
#> [1083,] -0.08715083 -2.012899916
#> [1084,]  0.40975797 -1.222276879
#> [1085,] -0.02162801 -0.771730742
#> [1086,]  0.35701878  1.758718834
#> [1087,]  0.78840476  1.984815672
#> [1088,]  0.24307053  1.711141430
#> [1089,]  0.35701878 -0.563903509
#> [1090,]  0.78840476  0.807590991
#> [1091,]  0.78840476  2.032393077
#> [1092,]  0.78840476  0.424119314
#> [1093,] -0.18831545  0.288368005
#> [1094,]  0.40975797 -1.590883142
#> [1095,] -0.13557627 -0.768520582
#> [1096,]  0.78840476 -0.658494499
#> [1097,] -1.16740751 -1.039471950
#> [1098,]  0.24307053  1.257517720
#> [1099,] -0.13557627 -1.949379280
#> [1100,]  0.78840476  1.122219167
#> [1101,] -0.13557627  1.146117895
#> [1102,] -2.04296310 -2.123545201
#> [1103,] -1.54605431 -1.153818757
#> [1104,] -0.13557627  0.486212603
#> [1105,] -1.11898206 -2.123545201
#> [1106,] -1.06624288 -1.258713128
#> [1107,] -0.56696225 -1.153818757
#> [1108,]  0.29149597  1.253156963
#> [1109,] -2.04296310 -2.539205084
#> [1110,]  0.78840476  2.050818932
#> [1111,] -1.06387104 -1.027140796
#> [1112,] -1.06387104 -2.539205084
#> [1113,]  0.78840476 -0.158272789
#> [1114,] -0.56696225 -1.111534207
#> [1115,]  0.78840476  0.223018007
#> [1116,]  0.78840476  3.319415193
#> [1117,]  0.78840476  2.587756484
#> [1118,]  0.78840476  0.690662601
#> [1119,]  0.24307053 -0.237937163
#> [1120,]  0.35701878  1.319160223
#> [1121,]  0.78840476  0.503198687
#> [1122,]  0.78840476  2.174378718
#> [1123,]  0.78840476  1.928277952
#> [1124,]  0.24307053  2.124519197
#> [1125,] -0.25383827  2.917820408
#> [1126,]  0.24307053  0.648464739
#> [1127,] -0.13989001  0.314624626
#> [1128,]  0.40975797 -0.219244728
#> [1129,]  0.24307053  1.474624995
#> [1130,]  0.78840476 -0.298510818
#> [1131,] -0.13989001  0.743646498
#> [1132,] -2.04296310 -0.866032110
#> [1133,]  0.40975797 -0.352019808
#> [1134,] -2.04296310 -2.125827318
#> [1135,] -0.56696225 -1.009598958
#> [1136,]  0.78840476  3.319415193
#> [1137,] -2.04296310 -0.669790865
#> [1138,] -0.13989001  1.829321653
#> [1139,]  0.78840476  1.978137473
#> [1140,]  0.78840476  3.319415193
#> [1141,] -0.51853681 -2.539205084
#> [1142,]  0.78840476  0.473881448
#> [1143,]  0.35701878  0.677631636
#> [1144,] -0.02162801 -1.730910732
#> [1145,] -1.06624288 -2.117188311
#> [1146,] -1.06624288 -2.117188311
#> [1147,] -0.02162801 -1.035779803
#> [1148,] -0.62207327 -1.323887099
#> [1149,] -1.54605431 -0.415540444
#> [1150,] -2.04296310 -1.130140081
#> [1151,] -0.13557627  0.670240613
#> [1152,]  0.40975797  0.618689133
#> [1153,]  0.78840476  2.032393077
#> [1154,] -0.02162801 -1.035779803
#> [1155,] -1.06624288 -2.117188311
#> [1156,] -0.02162801 -1.730910732
#> [1157,] -2.04296310 -1.130140081
#> [1158,] -0.13557627  0.670240613
#> [1159,] -0.51853681 -0.998251436
#> [1160,] -0.08715083  0.618689133
#> [1161,] -0.13557627 -0.768520582
#> [1162,] -1.06624288 -2.117188311
#> [1163,]  0.35701878  1.551032681
#> [1164,]  0.78840476  1.672423064
#> [1165,] -0.08715083  0.329617767
#> [1166,]  0.35701878  0.768699122
#> [1167,]  0.78840476 -0.515661008
#> [1168,]  0.40975797 -1.087803895
#> [1169,]  0.40975797 -0.063771242
#> [1170,]  0.78840476  1.168893030
#> [1171,] -0.13557627  0.732191956
#> [1172,]  0.24307053  0.243321996
#> [1173,]  0.40975797  0.861175756
#> [1174,] -1.06387104 -1.007414228
#> [1175,]  0.78840476  2.063945542
#> [1176,]  0.24307053  1.059398342
#> [1177,]  0.78840476 -1.597240032
#> [1178,] -0.13557627 -0.089904781
#> [1179,] -0.13989001  1.605123482
#> [1180,] -0.51853681  0.930002494
#> [1181,] -0.51853681  0.930002494
#> [1182,] -2.04296310 -2.015394142
#> [1183,]  0.35701878 -2.539205084
#> [1184,]  0.24307053  0.744523110
#> [1185,] -0.13557627  0.708602723
#> 
#> $scores$rot.scores
#> NULL
#> 
#> 
#> $loadings
#> Loading Method: BRM
#> 
#>       1     2    
#> wmt2  0.384      
#> wmt1  0.254      
#> wmt3  0.217 0.131
#> wmt5  0.201 0.142
#> wmt4  0.188  0.14
#> wmt9        0.293
#> wmt7        0.258
#> wmt15       0.247
#> wmt14       0.243
#> wmt6   0.14  0.24
#> wmt16        0.22
#> wmt8        0.219
#> wmt10 0.166 0.206
#> wmt12       0.187
#> wmt18       0.183
#> wmt13       0.171
#> wmt17       0.168
#> wmt11       0.155
#>